### 2021-02-06 #3015 @kyleniemeyer
Hi @D-K-E, thanks for your interest in JOSS. We just this pre-review issue to assign the editor and reviewers, and resolve any issues.

First, regarding the compilation error in your paper, I think you need `---` after the YAML header, before the `# Summary` line.

Second, your paper needs a bit of work. For one thing, I do not think that the summary is actually written for a non-specialist audience, because reading it I am not really sure what the purpose/functionality is. Also, the rest of the paper should not be part of the Statement of Need—that is typically around one paragraph. You can have other sections beyond Summary and Statement of Need, to describe the functionality, uses, etc. of the software. You should also connect the software to some theory, via citations to relevant articles.
