### 2021-02-06 #3015 @D-K-E
Hi @kyleniemeyer thanks for the input. I'll try to fix the compilation error. It compiled fine under the web service, but I'll add the line.
I'll clarify the points you had pointed out as well.
How should I inform JOSS once the changes are made ?
