+++ gmodels/gtypes/digraph.py
@@ -1,27 +1,45 @@
 """!
-Directed graph
+\file digraph.py
+
+# Directed Graph
+
+This is basically a Graph with directed edges. This has implications on several
+algorithms. Most notably connected components algorithm does not work the same
+way. The minimum spanning tree also does not work. We need to differentiate the
+positioning of nodes in edges as well. Sometimes it is also represented as \f[
+G = (V, A) \f] where A is arcs.
+
+Whenever possible we simply change the edge generation strategy in order to use
+the parent's algorithm.
+
 """
-from typing import Set, Optional, List, Tuple, Dict
+from typing import Set, Callable
 from gmodels.gtypes.edge import Edge
 from gmodels.gtypes.node import Node
-from gmodels.gtypes.path import Path, Cycle
-from gmodels.gtypes.tree import Tree
 from gmodels.gtypes.abstractobj import EdgeType
 from gmodels.gtypes.graph import Graph
 from gmodels.gtypes.undigraph import UndiGraph
 from uuid import uuid4
-import math
 
 
 class DiGraph(Graph):
     """!
-    Directed graph implementation
+    \brief Directed graph implementation
     """
 
     def __init__(
         self, gid: str, data={}, nodes: Set[Node] = None, edges: Set[Edge] = None
     ):
-        ""
+        """!
+        \brief Constructor for DiGraph
+
+        More or less what we have in Graph.
+        We just make sure every edge is a directed edge.
+        \sa Graph for parameters.
+
+        \throws ValueError if there is any undirected edge among the argument
+        edge set.
+        """
 
         if edges is not None:
             for edge in edges:
@@ -37,14 +55,33 @@ class DiGraph(Graph):
 
     @classmethod
     def from_graph(cls, g: Graph):
-        ""
+        """!
+        \brief make DiGraph from Graph
+
+        \param g argument graph
+
+        We give a random id for the resulting DiGraph.
+        """
         return DiGraph(
             gid=str(uuid4()), data=g.data(), nodes=g.nodes(), edges=g.edges()
         )
 
     def is_family_of(self, src: Node, dst: Node) -> bool:
         """!
-        Check if src is family of dst
+        \brief Check if src is family of dst
+
+        \todo What if the graph has self loops. What if src and dst are same
+        nodes.
+
+        Checks two positional conditions with respect to nodes.
+        This function can also be implemented as a one liner with something
+        like:
+        \code
+        return self.is_parent_of(n1,n2) or self.is_child_of(n1,n2)
+        \endcode
+        The current implementation is a little more efficient than that. There
+        is also a room for improvement, since it can be much more efficient
+        using edge list representation.
         """
         for e in self.edges():
             # dst is child of src
@@ -56,6 +93,14 @@ class DiGraph(Graph):
         return False
 
     def is_parent_of(self, parent: Node, child: Node) -> bool:
+        """!
+        \brief check if one node is parent of another node
+
+        We define the notion of parent node as the following.
+        For all e in E[G] and for all {v,w} in V[e] if e is an outgoing edge of
+        v and incoming edge of w than v is parent of w.
+        """
+
         def cond(n_1: Node, n_2: Node, e: Edge):
             ""
             c = n_1 == e.start() and e.end() == n_2
@@ -64,50 +109,90 @@ class DiGraph(Graph):
         return self.is_related_to(n1=parent, n2=child, condition=cond)
 
     def is_child_of(self, child: Node, parent: Node) -> bool:
-        ""
+        """!
+        \brief  check if one node is child of another node
+
+        We define the notion of child node as the following.
+        For all e in E[G] and for all {v,w} in V[e] if e is an incoming edge of
+        v and outgoing edge of w than v is child of w.
+        As you can see from the definition provided in is_parent_of() as well,
[…]
