+++ paper/paper.bib
@@ -1,35 +1,51 @@
- @book{Cowell_2005, place={New York}, title={Probabilistic networks and expert systems}, url={http://accesbib.uqam.ca/cgi-bin/bduqam/transit.pl?&noMan=25126878}, publisher={Springer-Verlag}, author={Cowell, Robert G}, year={2005} }
+ 
+ @inproceedings{Ankan_Panda_2015a, title={pgmpy: Probabilistic graphical models using python}, booktitle={Proceedings of the 14th python in science conference (SCIPY 2015)}, author={Ankan, Ankur and Panda, Abinash}, year={2015} }
 
-
- @article{Drton_2009, title={Discrete chain graph models}, volume={15}, ISSN={1350-7265}, url={http://arxiv.org/abs/0909.0843}, DOI={10.3150/08-BEJ172}, note={arXiv: 0909.0843}, number={3}, journal={Bernoulli}, author={Drton, Mathias}, year={2009}, month={Aug}, pages={736–753} }
-
-
- @book{Koller_Friedman_2009, place={Cambridge, MA}, series={Adaptive computation and machine learning}, title={Probabilistic graphical models: principles and techniques}, ISBN={978-0-262-01319-2}, publisher={MIT Press}, author={Koller, Daphne and Friedman, Nir}, year={2009}, collection={Adaptive computation and machine learning} }
-
-
- @book{Darwiche_2009, place={Cambridge ; New York}, title={Modeling and reasoning with Bayesian networks}, ISBN={978-0-521-88438-9}, publisher={Cambridge University Press}, author={Darwiche, Adnan}, year={2009} }
-
-
- @book{Erciyes_2018, place={Cham}, edition={1st ed. 2018}, series={Texts in Computer Science}, title={Guide to Graph Algorithms: Sequential, Parallel and Distributed}, ISBN={978-3-319-73235-0}, DOI={10.1007/978-3-319-73235-0}, publisher={Springer International Publishing : Imprint: Springer}, author={Erciyes, K.}, year={2018}, collection={Texts in Computer Science} }
-
- @book{Even_Guy Even_2012, place={Cambridge, NY}, edition={2nd ed}, title={Graph algorithms}, ISBN={978-0-521-51718-8}, publisher={Cambridge University Press}, author={Even, Shimon and Guy Even}, year={2012} }
-
+ @book{Ankan_Panda_2015b, title={Mastering probabilistic graphical models using python: master probabilistic graphical models by learning through real-world problems and illustrative code examples in Python}, ISBN={978-1-78439-468-4}, author={Ankan, Ankur and Panda, Abinash}, year={2015} }
 
  @book{Cohen_Joyner_Nguyen_2011, title={Algorithmic Graph Theory}, url={https://code.google.com/archive/p/graphbook/}, author={Cohen, Nathan and Joyner, David and Nguyen, Minh}, year={2011}, month={May} }
 
+ @book{Cowell_2005, place={New York}, title={Probabilistic networks and expert systems}, url={http://accesbib.uqam.ca/cgi-bin/bduqam/transit.pl?&noMan=25126878}, publisher={Springer-Verlag}, author={Cowell, Robert G}, year={2005} }
+
+ @book{Darwiche_2009, place={Cambridge ; New York}, title={Modeling and reasoning with Bayesian networks}, ISBN={978-0-521-88438-9}, publisher={Cambridge University Press}, author={Darwiche, Adnan}, year={2009} }
+
+ @book{Dechter_2019, title={Reasoning with probabilistic and deterministic graphical models: exact algorithms}, ISBN={978-1-68173-492-7}, author={Dechter, Rina}, year={2019} }
+
[…]
