+++ paper/paper.md
@@ -1,9 +1,9 @@
 ---
+
 title: "PyGModels: A Python package for exploring Probabilistic Graphical
 Models with Graph Theoretical Structures"
 
 tags:
-
 - Python
 - probabilistic graphical models
 - Bayesian statistics
@@ -34,8 +34,8 @@ algorithms that depend graph theoretical structures in order to compute
 probabilities in the case of PGMs. We do this by making the inferable
 probabilistic graph, a true graph in the sense of graph theory. We also
 implement other more specific PGMs like Bayesian Networks and Markov Random
-Fields again as a true digraph, and undirected graph, in the sense of graph
-theory. In most cases, the input of our library is, a set of random variables
+Fields again as a true digraph, and undirected graph, respectively. In the
+most common use cases, the input of our library is, a set of random variables
 (functions specified by a certain probability distribution), edges that relate
 these random variables to one another, and factors in the form of objects
 which are instantiated by a set of random variables and a real valued
@@ -47,44 +47,93 @@ theory or statistics.
 
 # Statement of Need
 
-Let us try to demonstrate the need for this library by a use case.
+Let us try to demonstrate the need for `PyGModels` by a use case. One has a
+set of categorical random variables in the form of a function specified by a
+probability distribution. One has a set of edges that encode a certain
+independence assumption over her random variables, and one has a set of
+factors, factorizes a certain probability distribution over her entire graph.
+Given these, the API of `PyGModels` might solve two major issues for the
+researcher or the student alike:
 
-Though the name suggests otherwise PGMs and related software, for example
+- Compute posterior probability distribution or most probable explanation
+  given certain evidence.
 
-- [pyGM](https://github.com/ihler/pyGM)
+- Test new algorithms of inference.
 
-- [pgmPy](https://github.com/indapa/pgmPy)
+Though the first issue is not irrelevant, the forte of `PyGModels` is the
+second issue due to its lightweight nature and its direct embodiment of
+statistical and graph theoretic [we follow mostly @Diestel_2017 for graph
+theoretic conventions and definitions; most of the graph algorithms come from
+@Erciyes_2018 and @Even_Guy; exact pages are cited in doc strings of related
+functions inside the source code] considerations in the same base class. The
+entire library depends only on python standard library which makes it very
+extendible and easy to integrate and adapt to other projects as well. Through
+its rigorous adoption of mathematical definitions of involved concepts, it
+becomes feasible to extend arbitrary factors through their point wise product,
+or apply common graph analysis algorithms such as finding articulation points
+or bridges, or finding an optimal path defined by a cost function.
 
-- [pgm](https://github.com/paulorauber/pgm)
+# Applications and Similar Works
 
-are in most cases closer in nature to statistics than to graph theory. This
-results in using specific structures like Factor Graphs, which do not
-necessarily conform to the definition of a graph in the sense of Graph Theory.
-The idea behind using these structures is that one has an established
-inference scheme, and these structures simply make it easier to apply
-inference algorithms. Our software takes the opposite approach. We privilege
-graph theoretical structures and extend them to accommodate other structures
-like factors.
+PGMs are known for their wide range of applications in computer vision,
+information retrieval, disease diagnosis and more recently, in the context of
+our phd thesis, annotations of ancient documents.
 
-Another important aspect of our library is its rigor in following mathematical
-definitions where possible of the involved concepts. In most text books,
-structures like factors and probability distributions are defined as but they
-are parametrized as tables in most cases. We define both distributions and
-factors as functions, and classes respectively. The factor is defined as a
[…]
