+++ paper/paper.bib
@@ -20,7 +20,7 @@
  @book{Erciyes_2018, place={Cham}, edition={1st ed. 2018}, series={Texts in Computer Science}, title={Guide to Graph Algorithms: Sequential, Parallel and Distributed}, ISBN={978-3-319-73235-0}, DOI={10.1007/978-3-319-73235-0}, publisher={Springer International Publishing : Imprint: Springer}, author={Erciyes, K.}, year={2018}, collection={Texts in Computer Science} }
  @inbook{Evans_2018, place={Boca Raton, Florida : CRC Press, c2019.}, edition={1}, title={Markov Properties for Mixed Graphical Models}, ISBN={978-0-429-46397-6}, url={https://www.taylorfrancis.com/books/9780429874246/chapters/10.1201/9780429463976-2}, DOI={10.1201/9780429463976-2}, booktitle={Handbook of Graphical Models}, publisher={CRC Press}, author={Evans, Robin}, editor={Maathuis, Marloes and Drton, Mathias and Lauritzen, Steffen and Wainwright, Martin}, year={2018}, month={Nov}, pages={39–60} }
 
- @book{Even_Guy Even_2012, place={Cambridge, NY}, edition={2nd ed}, title={Graph algorithms}, ISBN={978-0-521-51718-8}, publisher={Cambridge University Press}, author={Even, Shimon and Guy Even}, year={2012} }
+ @book{Even_Guy_Even_2012, place={Cambridge, NY}, edition={2nd ed}, title={Graph algorithms}, ISBN={978-0-521-51718-8}, publisher={Cambridge University Press}, author={Even, Shimon and Guy Even}, year={2012} }
 
  @book{Ihler_2020, title={pyGM}, url={https://github.com/ihler/pyGM}, author={Ihler, Alexander}, year={2020}, month={Oct} }
 
