+++ paper/paper.bib
@@ -48,4 +48,4 @@
 
  @book{Sucar_2015, place={London Heidelberg New York Dordrecht}, series={Advances in computer vision and pattern recognition}, title={Probabilistic graphical models: principles and applications}, ISBN={978-1-4471-6698-6}, publisher={Springer}, author={Sucar, Luis Enrique}, year={2015}, collection={Advances in computer vision and pattern recognition} }
 
- @book{Marloes Maathuis_Drton_Lauritzen_Wainwright_2019, place={Boca Raton, Florida}, title={Handbook of graphical models}, ISBN={978-0-429-87423-9}, publisher={CRC Press/Taylor and Francis Group}, year={2019} }
+ @book{Marloes_Maathuis_Drton_Lauritzen_Wainwright_2019, place={Boca Raton, Florida}, title={Handbook of graphical models}, ISBN={978-0-429-87423-9}, publisher={CRC Press/Taylor and Francis Group}, year={2019} }
