### 2021-02-09 #3015 @D-K-E
@dfm @kyleniemeyer I have made couple of changes to the paper as per requested. Here is its summary:

- Paper compiles now.
- Added `References` section for citations.
- Reworded `Summary` to make it more clear to general public.
-  `Statement of Need` tries to demonstrate the research related aspect through a use case and citations.
- Added `Applications and Similar Works` section which contains a very small review of similar libraries and how `PyGModels` compares against them. During the comparison, I also cite relevant sources when I am introducing a novelty with respect compared libraries.

A simple question, I cite some sources in the source code but not in the paper should I include them to `References` section as well ?
