### 2021-02-09 #3015 @kyleniemeyer
Thanks @D-K-E. The paper is definitely improved, and I'm going to leave it to @dfm now (aside from the following comments 😅).

I do think it makes sense to cite things in the paper that you cite in the source code—this is actually a great way of acknowledging those sources to give them the appropriate academic credit.

The Summary (and rest of the paper) is better, but I would suggest trying to make the introduction even more general. As written, it still does not quite satisfy the requirement (in my view):
> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
