### 2021-02-10 99c6588
factor documentation added small corrections on paper
