+++ gmodels/factor.py
@@ -25,6 +25,16 @@ class Factor(GraphObject):
         data={},
     ):
         """!
+        \brief Constructor for a factor \f[ \phi(A,B) \f]
+
+        A factor is defined by a set of random variables which constitutes its
+        scope and a real valued function. The canonical definition can be found
+        in Koller and Friedman 2009, p. 106-107.  During construction, we check
+        so that all random variables are positive valued. If no function is
+        provided, we instantiate the factor with a marginal joint function.
+
+        \param scope_vars variables that constitue scope of factor.
+        \param factor_fn a real valued function
         """
         # check all values are positive
         super().__init__(oid=gid, odata=data)
@@ -45,13 +55,39 @@ class Factor(GraphObject):
 
         self.Z = self.zval()
 
-    def scope_vars(self, f=lambda x: x):
+    def scope_vars(self, f=lambda x: x) -> Set[NumCatRVariable]:
+        """!
+        \brief get variables that are inside the scope of this factor
+
+        \param f is a function that transforms the scope of this factor.
+
+        \code
+
+        >>> A = NumCatRVariable("A",
+        >>>                     input_data={"outcome-values": [True, False]},
+        >>>                     distribution=lambda x: 0.5)
+        >>> fac = Factor(gid=str(uuid4()), scope_vars=set([A]))
+        >>> fac.scope_vars(f=lambda x: set([(x,x)]))
+        >>> set([(A, A)])
+
+        \endcode
+        """
         return f(self.svars)
 
     @classmethod
     def from_joint_vars(cls, svars: Set[NumCatRVariable]):
         """!
-        Make factor from joint variables
+        \brief Make factor from joint variables
+
+        \code
+
+        >>> A = NumCatRVariable("A",
+        >>>                     input_data={"outcome-values": [True, False]},
+        >>>                     distribution=lambda x: 0.5)
+
+        >>> fac = Factor.from_joint_vars(svars=set([A]))
+
+        \endcode
         """
         return Factor(gid=str(uuid4()), scope_vars=svars)
 
@@ -63,7 +99,29 @@ class Factor(GraphObject):
         fn: Optional[Callable[[Set[Tuple[str, NumericValue]]], float]] = None,
     ):
         """!
-        Make factor from joint variables
+        \brief Make factor from joint variables
+
+        \param X_i main variable
+        \param Pa_Xi parent variables of the main variable
+        \param fn factor function that defines the factor between parent and child
+
+        \code
+
+        >>> A = NumCatRVariable("A",
+        >>>                     input_data={"outcome-values": [True, False]},
+        >>>                     distribution=lambda x: 0.5)
+
+        >>> B = NumCatRVariable("B",
+        >>>                     input_data={"outcome-values": [True, False]},
+        >>>                     distribution=lambda x: 0.5)
+
+        >>> C = NumCatRVariable("C",
+        >>>                     input_data={"outcome-values": [True, False]},
+        >>>                     distribution=lambda x: 0.5)
+
+        >>> fac = Factor.from_conditional_vars(X_i=A, Pa_Xi=set([B,C]))
+
+        \endcode
         """
         Pa_Xs = 1.0
         for p in Pa_Xi:
@@ -91,7 +149,12 @@ class Factor(GraphObject):
         value_transform=lambda x: x,
     ) -> List[Set[Tuple[str, NumericValue]]]:
         """!
-        Get factor domain Val(D) D being a set of random variables
+        \brief Get factor domain Val(D) D being a set of random variables
+
+        \param D set of random variables
+        \param rvar_filter filtering function for random variables
+        \param value_filter filtering values from random variables' codomain
[…]
