### 2021-03-06 #3015 @dfm
@D-K-E: Thanks for your patience! The paper is looking better, but I think that the online documentation could still use some work to make it more accessible. Right now the README seems to be the main source of documentation with the doxygen API docs hosted separately. It would be good to connect these (probably by moving most of the functionality discussion to the doxygen pages, but feel free to structure it differently) and add more context because right now it's very hard to navigate and see how the different pieces fit together.

Also, can you explain the relationship between this repo and https://github.com/Viva-Lambda/graphical-models?
