### 2021-03-06 #3015 @D-K-E
@dfm Hi ! It is really nice to here from you. The Viva-Lambda repository, is also me, a part of the code is live coded in my twitch channel (twitch.tv/vivalambda) and some of my viewers, mostly graduate students in CS, were interested in seeing the code, so I uploaded it to channel's repository as well. I can remove it, if you feel it's necessary, or add a notice to readme which repo is the original one. I tend to use the channel's repository as a scratch book (test new ideas, functionality, discover new apis, etc), and use my personal repository, D-K-E to publish more stable versions.

For the documentation, I will move the "Currently Supported Models" to doxygen main page, and continue to add doc strings to functions. Does that sound alright ?
