### 2021-03-13 #3015 @D-K-E
@dfm I have moved usage examples to doxygen main page as I understood from:

>  It would be good to connect these (probably by moving most of the functionality discussion to the doxygen pages, but feel free to structure it differently)

I'll see what I can do about adding more context. Would you mind giving an example repository who does this well ? I am not sure how to approach it
