### 2021-03-15 #3015 @dfm
@D-K-E:

- Thanks for the clarification re: Viva-Lambda; I think it would be good to add a link to that README highlighting that the D-K-E version is the primary repo and clarifying the relationship, but otherwise all sounds good!
- I think that moving the README content to the doxygen pages really helps and it's getting closer. I'd say that the main major component that is still missing is the [Example Usage](https://joss.readthedocs.io/en/latest/review_criteria.html#example-usage) review criterion. The one example is good, but I think it would be useful to have more extensive examples. The [pgmpy](https://pgmpy.org/) docs (for example) are a pretty good example of what we're looking for. Yours certainly does not need to be quite so extensive, but those docs are a good example of the structure, I think.

I'll start looking for reviewers while you chip away at this. Let me know if you have further questions!
