### 2021-03-16 #3115 @whedon
[REVIEW]: PyGModels: A Python package for exploring Probabilistic Graphical Models with Graph Theoretical Structures
**Submitting author:** @D-K-E (<a href="http://orcid.org/0000-0002-1552-8938">D. Kaan Eraslan</a>)
**Repository:** <a href="https://github.com/D-K-E/graphical-models/" target ="_blank">https://github.com/D-K-E/graphical-models/</a>
**Version:** v0.1.0
