### 2021-03-17 #3115 @eigenfoo
Hello! I'm still in the middle of my review, but have one question so far.

From the [README](https://github.com/D-K-E/graphical-models/):

> The primary goal is to facilitate the understanding of models and basic inference strategies using well documented data structures based only on Python 3 standard library.

> As the overall library is not built for efficiency, we recommend not to use it in production. It should not be to difficult to transfer the concepts introduced in the source code though.

It seems that this repository is mainly written for pedagogical purposes, and perhaps not for actual use in scientific applications - in other words, it sounds like the library is meant to be read, and not run. @D-K-E am I correct in this? If so, @dfm can you advise if such a library is in scope for JOSS publication? I'm unsure if this library satisfies the "research software" requirement of JOSS - some clarity would be helpful.

---

Also noting from this thread (https://github.com/openjournals/joss-reviews/issues/3015#issuecomment-792050736) that @Viva-Lambda is another account for @D-K-E, so the target repo definitely passes the "contribution and authorship" check box.
