### 2021-03-17 D-K-E/graphical-models#1 @eigenfoo
Unable to create conda virtual environment, per installation instructions
[ cc openjournals/joss-reviews/issues/3115 ]

I'm unable to create a conda environment using the [installation instructions found on the README](https://github.com/D-K-E/graphical-models#installation). However, I'm able to simply `pip install .` the package, so there's one viable path to installation! I've included the error message and details below.

I imagine it's simply because of https://github.com/conda/conda/issues/7311#issuecomment-442320274 (or some similar issue), but can't confirm because I don't know what platform the `environment.yml` was exported on.

OS: macOS Catalina
Conda version: 4.8.3

```
(base) george@eigenfoo:~/graphical-models
$ conda env create -f environment.yml
Solving environment: failed

ResolvePackageNotFound:
  - python==3.6.12=hffdb5ce_0_cpython
  - _openmp_mutex==4.5=1_gnu
  - tk==8.6.10=h21135ba_1
  - xz==5.2.5=h516909a_1
  - openssl==1.1.1i=h7f98852_0
  - ca-certificates==2020.12.5=ha878542_0
  - libstdcxx-ng==9.3.0=h2ae2ef3_17
  - sqlite==3.34.0=h74cdb3f_0
  - readline==8.0=he28a2e2_2
  - certifi==2020.12.5=py36h5fab9bb_0
  - ld_impl_linux-64==2.35.1=hea4e1c9_1
  - libgomp==9.3.0=h5dbcf3e_17
  - _libgcc_mutex==0.1=conda_forge
  - libgcc-ng==9.3.0=h5dbcf3e_17
  - setuptools==49.6.0=py36h9880bd3_2
  - zlib==1.2.11=h516909a_1010
  - libffi==3.3=h58526e2_2
  - ncurses==6.2=h58526e2_4
 ```
