### 2021-03-17 D-K-E/graphical-models#1 @D-K-E
All right I'll change it to `pip install .` since it is all standard library anyway
