### 2021-03-17 #3115 @D-K-E
@dfm I am adding them right away.

@eigenfoo It is a little complicated issue. I am using this library for demonstrating how some data <sup>1</sup>  can be analyzed using probabilistic graphical models in my thesis so it is used at least in the case of one scientific work. But my focus during the writing of the library was to be as close as possible to definitions provided by text books (mostly Koller, Friedman 2009) so that there would be less friction between the text book and the implementation. The pedagogical aspect is more of an intended side effect which comes from being close to the text book. However the goal was to reproduce the definitions as close as possible **not** to teach them.

---
<sup>1</sup>  historical documents annotated with a certain flavor of RDFa format.
