+++ README.md
@@ -1,6 +1,6 @@
 # Graphical Models
 
-![Build Status](https://github.com/D-K-E/graphical-models/workflows/Python%20Package%20using%20Conda/badge.svg?branch=master)
+[![Python package](https://github.com/D-K-E/graphical-models/actions/workflows/python-package.yml/badge.svg?branch=master)](https://github.com/D-K-E/graphical-models/actions/workflows/python-package.yml)
 
 
 [![DOI](https://zenodo.org/badge/321839625.svg)](https://zenodo.org/badge/latestdoi/321839625)
