+++ paper/paper.bib
@@ -1,5 +1,8 @@
  
- @inproceedings{Ankan_Panda_2015a, title={pgmpy: Probabilistic graphical models using python}, booktitle={Proceedings of the 14th python in science conference (SCIPY 2015)}, author={Ankan, Ankur and Panda, Abinash}, year={2015} }
+ @inproceedings{Ankan_Panda_2015a, title={pgmpy: Probabilistic graphical
+     models using python}, booktitle={Proceedings of the 14th python in
+         science conference (SCIPY 2015)}, author={Ankan, Ankur and Panda,
+             Abinash}, doi={10.25080/Majora-7b98e3ed-001 }, year={2015} }
 
  @book{Ankan_Panda_2015b, title={Mastering probabilistic graphical models using python: master probabilistic graphical models by learning through real-world problems and illustrative code examples in Python}, ISBN={978-1-78439-468-4}, author={Ankan, Ankur and Panda, Abinash}, year={2015} }
 
@@ -9,13 +12,18 @@
 
  @book{Darwiche_2009, place={Cambridge ; New York}, title={Modeling and reasoning with Bayesian networks}, ISBN={978-0-521-88438-9}, publisher={Cambridge University Press}, author={Darwiche, Adnan}, year={2009} }
 
- @book{Dechter_2019, title={Reasoning with probabilistic and deterministic graphical models: exact algorithms}, ISBN={978-1-68173-492-7}, author={Dechter, Rina}, year={2019} }
+ @book{Dechter_2019, title={Reasoning with probabilistic and deterministic
+     graphical models: exact algorithms}, ISBN={978-1-68173-492-7},
+     author={Dechter, Rina}, doi={10.2200/S00893ED2V01Y201901AIM041}, year={2019} }
 
  @book{Diestel_2017, place={Hamburg}, edition={5}, title={Graph Theory}, ISBN={978-3-662-53621-6}, publisher={Springer}, author={Diestel, Reinhard}, year={2017} }
 
  @article{Drton_2009, title={Discrete chain graph models}, volume={15}, ISSN={1350-7265}, url={http://arxiv.org/abs/0909.0843}, DOI={10.3150/08-BEJ172}, note={arXiv: 0909.0843}, number={3}, journal={Bernoulli}, author={Drton, Mathias}, year={2009}, month={Aug}, pages={736–753} }
 
- @article{Eppstein_Löffler_Strash_2010, title={Listing All Maximal Cliques in Sparse Graphs in Near-optimal Time}, url={http://arxiv.org/abs/1006.5440}, note={arXiv: 1006.5440}, journal={arXiv:1006.5440 [cs]}, author={Eppstein, David and Löffler, Maarten and Strash, Darren}, year={2010}, month={Jun} }
+ @article{Eppstein_Löffler_Strash_2010, title={Listing All Maximal Cliques in
+     Sparse Graphs in Near-optimal Time},
+     url={http://arxiv.org/abs/1006.5440}, note={arXiv: 1006.5440},
+     journal={arXiv:1006.5440 [cs]}, doi={10.1007/978-3-642-17517-6_36}, author={Eppstein, David and Löffler, Maarten and Strash, Darren}, year={2010}, month={Jun} }
 
  @book{Erciyes_2018, place={Cham}, edition={1st ed. 2018}, series={Texts in Computer Science}, title={Guide to Graph Algorithms: Sequential, Parallel and Distributed}, ISBN={978-3-319-73235-0}, DOI={10.1007/978-3-319-73235-0}, publisher={Springer International Publishing : Imprint: Springer}, author={Erciyes, K.}, year={2018}, collection={Texts in Computer Science} }
[…]
