### 2021-03-17 #3115 @D-K-E
@dfm added the requested DOIs, plus some changes in the documentation. I'll continue to improve the docs.
