### 2021-03-17 #3115 @dfm
@eigenfoo: Great question! I am satisfied by your response @D-K-E, but the paper and documentation should make the research applications very clear and demonstrate where this package fits into the ecosystem. So perhaps the discussion in the README should be clarified with this in mind.
