### 2021-03-21 #3115 @eigenfoo
@D-K-E this is great work! Detailed comments below.

---

### Summary and Statements of Need

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

> Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

> A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

In both the documentation and the paper, the summary and statement are quite technical, does not seem to be for a non-specialist audience, and does not make it clear who the target audience is.

- **Summary is too technical:** what distinguishes this library from prior art is quite technical, but nevertheless I think it could be introduced in a friendlier way: I hope you don't mind if I break down point-by-point suggestions:
  1. PGMs are widely used
  2. However, most existing PGM libraries implement PGMs in a way that _ignores_ their graphical nature (e.g. as explained on lines 64-66, "a factor is specified through an array of values", instead of a set of random variables and a function).
  3. PyGModel's value proposition is that it faithfully implements the graphical nature of PGMs, thereby giving PyGModel's instantiated objects both graph-theoretical and statistical properties, which allows users to explore and test inference algorithms that are rooted in both graph theory or statistics (paraphrasing lines 19-21)
  * @D-K-E would you say that I've summarized the paper correctly?

- **Statements of need are unclear:** the statements may be more technical than the summary, but after reading them I am still unsure who the intended users/audience is. If I'm correct in my previous summary, then the intended users are _researchers exploring inference algorithms that depend on graph-theoretical properties of PGMs_. PyGModel's lightweight nature, easy extensibility and faithful implementations of textbook algorithms are all good points, but become a lot clearer once I know that this library is for researchers!

---

### Community Guidelines

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Right now the [README](https://github.com/D-K-E/graphical-models#guide-for-contributors) outlines how users might contribute to the software, but doesn't explicitly ask users who have issues/need support to file an issue. It's probably a safe bet from a user's perspective, but best to make this explicit!

---

### Citations

From the paper, on lines 44-46:

> PGMs are known for their wide range of applications in computer vision, information retrieval, disease diagnosis and more recently, in the context of our phd thesis, annotations of ancient documents.

It would be great to cite any portion of your PhD (or related work) that may already be published, _especially_ if it showcases "real-world usage" of PyGModels. I understand if this isn't possible or if you may be unwilling though, so I'll go ahead and cross off the "references" checkbox.

---

### Well Written

Overall, the paper is well-written: there are some minor grammatical mistakes (e.g. phd/PhD on line 45, depend/depends on line 62, not/no on line 67), but these errors don't get in the way of comprehension. I'm willing to let these errors slide, so I'll cross off the "quality of writing" checkbox.
