### 2021-03-21 D-K-E/graphical-models#2 @eigenfoo
Ask users with issues/need support to file a GitHub issue?
[cc openjournals/joss-reviews/issues/3115]

From the review checklist:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Right now the [README](https://github.com/D-K-E/graphical-models#guide-for-contributors) outlines how users might contribute to the software, but doesn't explicitly ask users who have issues or need support to file an issue on GitHub. It's probably a safe bet from a user's perspective, but best to make this explicit on the README and/or documentation!
