### 2021-03-22 cee19b0
readme change grammatical errors corrected
