+++ paper/paper.md
@@ -74,7 +74,7 @@ cost function.
 
 PGMs are known for their wide range of applications in computer vision,
 information retrieval, disease diagnosis and more recently, in the context of
-our phd thesis, annotations of ancient documents.
+our PhD thesis, annotations of ancient documents.
 
 Other open sourced python libraries about PGMs include the following:
 
@@ -103,12 +103,12 @@ rather than a dedicated library. We will make a small comparison with `pgmpy`
 most of our remarks hold for other alternatives as well.
 
 `PyGModels` distinguish from `pgmpy` by its lightweight nature (`PyGModels`
-depend only on python 3.6 standard library). Our test suit cites its source
+depends only on python 3.6 standard library). Our test suit cites its source
 for most of the compared values inside doc string of functions for key
 functions like inference over graphs. Factors are specified by a set of
 random variables and a function whose domain is the cartesian product of
 codomains of random variables. In all of the libraries above, a factor is
-specified through an array of values. This has not direct implications on
+specified through an array of values. This has no direct implications on
 the output. However, it has implications on the evaluation order of
 operations. Our implementation is lazier and it conforms to the definition
 provided by Koller and Friedman [see @Koller_Friedman_2009 p.  106-107].
