+++ gmodels/__init__.py
@@ -94,179 +94,78 @@ Ran 179 tests in 0.666s
 OK
 ```
 
+
 ## Guide for Contributors
 
 As of now, most of the functions are unit tested. The test suit contains
-around 160 tests covering most of the important functionality. However, there
-can never be too much tests, so feel free to PR some of your own.
+around 170 tests covering most of the important functionality. However, there
+can never be too much tests, so feel free to create a pull request with some
+of your own.
 
-Another area of improvement is documentation.
+Another area of improvement is documentation. As of now, we lack usage
+examples for graph theoretical functionality of the library. Some functions
+can also use more elaborate docstrings. Notice that we are not using sphinx or
+other regular pythonic documentation generators. We are using
+[doxygen](https://www.doxygen.nl/index.html). For the willing user, we provide
+two templates for filling out docstrings in case she is not familiar with
+doxygen way of doing things:
 
-Adding other inference strategies are also welcome.
+- Partial template:
 
-Just make an issue and we can discuss the rest.
+```python
 
+def my_function(myarg1: str, myarg2: int) -> str:
+    """!
+    \brief One line explanation of functionality
 
-## Currently Supported Models
+    \param myarg1 description of the argument
+    \param myarg2 description of the argument
 
-There are currently 4 primary models that might interest the general public:
-
-### PGModel
-
-`PGModel`, a graph with defined as `G=(V,E,F)` where `V` is vertex set
-composed of `NumCatRVariable`s which are numeric categorical random variables,
-`E` is edge set, and `F` is factor set.  
-If `F` is none the factor set is deduced from edge set where each edge is
-considered as a factor whose scope is the set of incident nodes to edge.
-Sum-Product and Max-Product variable elimination are supported by `PGModel`.
-Evidence encoded in nodes are directly used for reducing factors during the
-instantiation of `PGModel`.
-Additional evidence can be provided at query time as well for conditional and
-mpe queries. `Marginal Map` queries are not yet supported.
-
-Usage:
-\code{.py}
-
-from gmodels.pgmodel import PGModel
-from gmodels.gtypes.edge import Edge, EdgeType
-from gmodels.factor import Factor
-from gmodels.randomvariable import NumCatRVariable
-
-# Example adapted from Darwiche 2009, p. 140
-
-idata = {
-            "a": {"outcome-values": [True, False]},
-            "b": {"outcome-values": [True, False]},
-            "c": {"outcome-values": [True, False]},
-        }
-
-a = NumCatRVariable(
-            node_id="a", input_data=idata["a"], distribution=lambda x: 0.6 if x else 0.4
-        )
-b = NumCatRVariable(
-    node_id="b", input_data=idata["b"], distribution=lambda x: 0.5 if x else 0.5
-)
-c = NumCatRVariable(
-    node_id="c", input_data=idata["c"], distribution=lambda x: 0.5 if x else 0.5
-)
-ab = Edge(
-    edge_id="ab",
-    edge_type=EdgeType.UNDIRECTED,
-    start_node=a,
-    end_node=b,
-)
-bc = Edge(
-    edge_id="bc",
-    edge_type=EdgeType.UNDIRECTED,
-    start_node=b,
-    end_node=c,
-)
-
-def phi_ba(scope_product):
-    ""
-    ss = set(scope_product)
-    if ss == set([("a", True), ("b", True)]):
-        return 0.9
-    elif ss == set([("a", True), ("b", False)]):
-        return 0.1
[…]
