+++ README.md
@@ -95,14 +95,74 @@ OK
 ## Guide for Contributors
 
 As of now, most of the functions are unit tested. The test suit contains
-around 160 tests covering most of the important functionality. However, there
-can never be too much tests, so feel free to PR some of your own.
+around 170 tests covering most of the important functionality. However, there
+can never be too much tests, so feel free to create a pull request with some
+of your own.
 
-Another area of improvement is documentation.
+Another area of improvement is documentation. As of now, we lack usage
+examples for graph theoretical functionality of the library. Some functions
+can also use more elaborate docstrings. Notice that we are not using sphinx or
+other regular pythonic documentation generators. We are using
+[doxygen](https://www.doxygen.nl/index.html). For the willing user, we provide
+two templates for filling out docstrings in case she is not familiar with
+doxygen way of doing things:
 
-Adding other inference strategies are also welcome.
+- Partial template:
 
-Just make an issue and we can discuss the rest.
+```python
+
+def my_function(myarg1: str, myarg2: int) -> str:
+    """!
+    \brief One line explanation of functionality
+
+    \param myarg1 description of the argument
+    \param myarg2 description of the argument
+
+    \return description of the returned value
+    """
+    return myarg1 + str(myarg2)
+```
+
+- Full template:
+
+```python
+
+def my_function(myarg1: str, myarg2: int) -> str:
+    """!
+    \brief One line explanation of functionality
+
+    Long multilined
+    description of
+    functionality
+
+    \param myarg1 description of the argument
+    \param myarg2 description of the argument
+
+    \exception TypeError description of the exception 
+
+    \return description of the returned value
+
+    \code{.py}
+
+    >>> a = my_function("Lucky number is ", 7)
+    >>> print(a)
+    >>> "Lucky number is 7"
+
+    \endcode
+
+    """
+    if isinstance(myarg2, int) is False:
+        raise TypeError(
+            "myarg2 " + str(myarg2) + " is of type " +
+            str(type(myarg2))
+            )
+    return myarg1 + str(myarg2)
+```
+
+Besides adding a documentation, you can also add other inference strategies.
+
+Just file an issue in the case of doubt or just signal your intent and we can
+discuss the rest.
 
 ## Contributors
 
