+++ gmodels/__init__.py
@@ -112,7 +112,7 @@ doxygen way of doing things:
 
 - Partial template:
 
-```python
+\code{.py}
 
 def my_function(myarg1: str, myarg2: int) -> str:
     """!
@@ -124,11 +124,11 @@ def my_function(myarg1: str, myarg2: int) -> str:
     \return description of the returned value
     """
     return myarg1 + str(myarg2)
-```
+\endcode
 
 - Full template:
 
-```python
+\code{.py}
 
 def my_function(myarg1: str, myarg2: int) -> str:
     """!
@@ -160,7 +160,7 @@ def my_function(myarg1: str, myarg2: int) -> str:
             str(type(myarg2))
             )
     return myarg1 + str(myarg2)
-```
+\endcode
 
 Besides adding a documentation, you can also add other inference strategies.
 
