+++ gmodels/__init__.py
@@ -115,14 +115,14 @@ doxygen way of doing things:
 \code{.py}
 
 def my_function(myarg1: str, myarg2: int) -> str:
-    """!
+    \"\"\"!
     \brief One line explanation of functionality
 
     \param myarg1 description of the argument
     \param myarg2 description of the argument
 
     \return description of the returned value
-    """
+    \"\"\"
     return myarg1 + str(myarg2)
 \endcode
 
@@ -131,7 +131,7 @@ def my_function(myarg1: str, myarg2: int) -> str:
 \code{.py}
 
 def my_function(myarg1: str, myarg2: int) -> str:
-    """!
+    \"\"\"!
     \brief One line explanation of functionality
 
     Long multilined
@@ -153,7 +153,7 @@ def my_function(myarg1: str, myarg2: int) -> str:
 
     \endcode
 
-    """
+    \"\"\"
     if isinstance(myarg2, int) is False:
         raise TypeError(
             "myarg2 " + str(myarg2) + " is of type " +
@@ -170,6 +170,19 @@ discuss the rest.
 ## Contributors
 
 - [Nihan](https://github.com/comecloseridontbyte)
+
+
+## Citation
+
+This library has a mirror in another repository as well. The [Viva-Lambda
+repository](https://github.com/Viva-Lambda/graphical-models) is in principal
+identical to the
+[D-K-E/graphical-models](https://github.com/D-K-E/graphical-models) (it might
+be one or two commits behind). During the development of the library, the
+weekly effort had been committed to that repository as well. However since the
+DOI refers to D-K-E/graphical-models, it should be considered as the main
+repository, and any references should refer to that one and not to
+Viva-Lambda/graphical-models.
 
 
 """
