### 2021-03-23 4e62053
removed a word from readme and mainpage
