### 2021-03-23 D-K-E/graphical-models#2 @D-K-E
@eigenfoo I added two templates for helping users to add documentation, and explicitly mention filing an issue at the end. How does that look ?
