### 2021-03-23 #3115 @D-K-E
@eigenfoo Well, first of all thank you for the suggestions, they really helped me to better phrase the problem.

---
## Summary

I don't think I could do better than your suggestions for the summary. I tried to incorporate them directly to the paper. It significantly decreased the length of the summary paragraph.

## Statements of need

I added the following remark to the beginning of the section as per the request for the clarification of the intended public of the package:
>Though the students of computer science or statistics might find a pedagogical
>value going through source code along with a textbook on probabilistic
>graphical models (something like Sucar [see @Sucar_2015] or Cowell [see
>@Cowell_2005] or Koller and Friedman [see @Koller_Friedman_2009]), we believe
>that the value proposition of `PyGModels` speaks mostly to researchers.

@dfm I hope that this clarifies how the packaged fits into the ecosystem, and resolves the discussion we had about the pedagogical use of it. I can add more details, if you find  necessary.
---

## Community Guidelines

I provided two templates for filling the docstrings of functions and explicitly asked users to file an issue with their doubt or intent in the `README` file.
---

## Citations

Unfortunately published parts of my PhD does not involve the direct use of this library. I am hoping to present a paper at conference with results produced by this library, once I discuss things more with my supervisor. The problem is that there is not enough data (historical documents in the right format) to produce something immediately useful. Hence I need to either wait for experts to produce data, or find someway to transform already existing data to right format. In any case, I need to discuss it with my supervisor.
---

## Well Written

As you might have noticed by now, English is not my first language, so there might be grammatical errors, and some awkward phrases here and there. I corrected the ones you pointed out. Feel free to make other suggestions regarding the language as well.
