+++ gmodels/__init__.py
@@ -117,15 +117,16 @@ doxygen way of doing things:
 \code{.py}
 
 def my_function(myarg1: str, myarg2: int) -> str:
-    '''!
+    \"\"\"!
     \brief One line explanation of functionality
 
     \param myarg1 description of the argument
     \param myarg2 description of the argument
 
     \return description of the returned value
-    '''
+    \"\"\"
     return myarg1 + str(myarg2)
+
 \endcode
 
 - Full template:
@@ -133,7 +134,7 @@ def my_function(myarg1: str, myarg2: int) -> str:
 \code{.py}
 
 def my_function(myarg1: str, myarg2: int) -> str:
-    '''!
+    \"\"\"!
     \brief One line explanation of functionality
 
     Long multilined
@@ -155,13 +156,14 @@ def my_function(myarg1: str, myarg2: int) -> str:
 
     \endcode
 
-    '''
+    \"\"\"
     if isinstance(myarg2, int) is False:
         raise TypeError(
             "myarg2 " + str(myarg2) + " is of type " +
             str(type(myarg2))
             )
     return myarg1 + str(myarg2)
+
 \endcode
 
 Besides adding a documentation, you can also add other inference strategies.
