+++ gmodels/__init__.py
@@ -117,14 +117,14 @@ doxygen way of doing things:
 \verbatim
 
 def my_function(myarg1: str, myarg2: int) -> str:
-    """!
+    '''!
     \brief One line explanation of functionality
 
     \param myarg1 description of the argument
     \param myarg2 description of the argument
 
     \return description of the returned value
-    """
+    '''
     return myarg1 + str(myarg2)
 
 \endverbatim
@@ -135,7 +135,7 @@ def my_function(myarg1: str, myarg2: int) -> str:
 \verbatim
 
 def my_function(myarg1: str, myarg2: int) -> str:
-    """!
+    '''!
     \brief One line explanation of functionality
 
     Long multilined
@@ -157,7 +157,7 @@ def my_function(myarg1: str, myarg2: int) -> str:
 
     \endcode
 
-    """
+    '''
     if isinstance(myarg2, int) is False:
         raise TypeError(
             "myarg2 " + str(myarg2) + " is of type " +
