+++ gmodels/__init__.py
@@ -8,7 +8,7 @@
 
 [![DOI](https://zenodo.org/badge/321839625.svg)](https://zenodo.org/badge/latestdoi/321839625)
 
-See doxygen generated [documentation](https://d-k-e.github.io/graphical-models/)
+See doxygen generated [documentation](https://viva-lambda.github.io/graphical-models/)
 
 
 The source code of this library aims to be accessible to all those interested
@@ -96,7 +96,6 @@ Ran 179 tests in 0.666s
 OK
 ```
 
-
 ## Guide for Contributors
 
 As of now, most of the functions are unit tested. The test suit contains
@@ -114,28 +113,26 @@ doxygen way of doing things:
 
 - Partial template:
 
-\verbatim
+```python
 
 def my_function(myarg1: str, myarg2: int) -> str:
-    '''!
+    \"\"\"!
     \brief One line explanation of functionality
 
     \param myarg1 description of the argument
     \param myarg2 description of the argument
 
     \return description of the returned value
-    '''
+    \"\"\"
     return myarg1 + str(myarg2)
-
-\endverbatim
-
+```
 
 - Full template:
 
-\verbatim
+```python
 
 def my_function(myarg1: str, myarg2: int) -> str:
-    '''!
+    \"\"\"!
     \brief One line explanation of functionality
 
     Long multilined
@@ -157,7 +154,7 @@ def my_function(myarg1: str, myarg2: int) -> str:
 
     \endcode
 
-    '''
+    \"\"\"
     if isinstance(myarg2, int) is False:
         raise TypeError(
             "myarg2 " + str(myarg2) + " is of type " +
@@ -165,28 +162,178 @@ def my_function(myarg1: str, myarg2: int) -> str:
             )
     return myarg1 + str(myarg2)
 
-\endverbatim
+```
 
 Besides adding a documentation, you can also add other inference strategies.
 
 Just file an issue in the case of doubt or signal your intent and we can
 discuss the rest.
 
+
+## Currently Supported Models
+
+There are currently 4 primary models that might interest the general public:
+
+### PGModel
+
+`PGModel`, a graph with defined as `G=(V,E,F)` where `V` is vertex set
+composed of `NumCatRVariable`s which are numeric categorical random variables,
+`E` is edge set, and `F` is factor set.  
+If `F` is none the factor set is deduced from edge set where each edge is
+considered as a factor whose scope is the set of incident nodes to edge.
+Sum-Product and Max-Product variable elimination are supported by `PGModel`.
+Evidence encoded in nodes are directly used for reducing factors during the
+instantiation of `PGModel`.
+Additional evidence can be provided at query time as well for conditional and
+mpe queries. `Marginal Map` queries are not yet supported.
+
+Usage:
+\code{.py}
+
+from gmodels.pgmodel import PGModel
+from gmodels.gtypes.edge import Edge, EdgeType
+from gmodels.factor import Factor
+from gmodels.randomvariable import NumCatRVariable
+
+# Example adapted from Darwiche 2009, p. 140
+
+idata = {
+            "a": {"outcome-values": [True, False]},
+            "b": {"outcome-values": [True, False]},
+            "c": {"outcome-values": [True, False]},
+        }
+
+a = NumCatRVariable(
+            node_id="a", input_data=idata["a"], distribution=lambda x: 0.6 if x else 0.4
+        )
+b = NumCatRVariable(
+    node_id="b", input_data=idata["b"], distribution=lambda x: 0.5 if x else 0.5
+)
+c = NumCatRVariable(
[…]
