### 2021-03-29 D-K-E/graphical-models#2 @eigenfoo
Great, thanks for adding!
