### 2021-03-31 D-K-E/graphical-models#3 @ankurankan
Missing code examples in documentation
In the documentation at https://d-k-e.github.io/graphical-models/html/index.html, I can see an example of PGModel but for other model types, there are no code examples. Also, it would be nice if you could use a real dataset/model for the examples, but if you are using a toy example, it would be helpful if you could describe it a bit before the code.

Ref: https://github.com/openjournals/joss-reviews/issues/3115
