### 2021-03-31 D-K-E/graphical-models#4 @ankurankan
Documentation Suggestions
I think the documentation still needs quite a bit of work.

1. I would really suggest adding more information to the docstrings: expected parameters, examples for using that method etc. I also think sphinx might be a much better option compared to doxygen for python, just in case you haven't checked it out.
2. I think the method descriptions can be improved as well. Currently, I see a lot of descriptions just pointing to the reference but I think it would be useful if there is a short description in the docstring as well as the reference for more details.

Ref: https://github.com/openjournals/joss-reviews/issues/3115
