### 2021-03-31 D-K-E/graphical-models#5 @ankurankan
Contributing guide
I see that there is some information for contributors in the readme. But it would be nice to add a github contribution guide as well at: https://github.com/D-K-E/graphical-models/community.

Ref: https://github.com/openjournals/joss-reviews/issues/3115
