### 2021-04-02 2b58ab7
added factor documentation
