### 2021-04-05 db75e89
added more documentation to factors
