### 2021-04-05 c2a02be
added new tests and documentation for factor
