### 2021-04-05 27ebbfe
added new tests and documentation for factor
