+++ gmodels/factor.py
@@ -98,9 +98,11 @@ class Factor(GraphObject):
         else:
             self.factor_fn = factor_fn
 
-        self.scope_products: List[Set[Tuple[str, NumericValue]]] = []
+        self.scope_products: List[Set[Tuple[str, NumericValue]]] = list(
+            product(*self.vars_domain())
+        )
 
-        self.Z = self.zval()
+        self.Z = self.partition_value(self.vars_domain())
 
     def scope_vars(self, f=lambda x: x) -> Set[NumCatRVariable]:
         """!
@@ -299,7 +301,7 @@ class Factor(GraphObject):
         rvar_filter=lambda x: True,
         value_filter=lambda x: True,
         value_transform=lambda x: x,
-    ):
+    ) -> list:
         """!
         \brief \see Factor.matches(rvar_filter, value_filter, value_transform)
         For a factor phi(A,B) return factor function's domain values, such as:
@@ -339,6 +341,7 @@ class Factor(GraphObject):
     ) -> Set[NumCatRVariable]:
         """!
         \brief Given a domain of values obtain scope variables implied
+
         Obtain random variables from given factor domain.
         Each value in domain comes with an identifier of its random variable.
         From these identifiers, we obtain set of random variables attested in
@@ -420,9 +423,13 @@ class Factor(GraphObject):
         """
         return self.normalize(self.phi(scope_product))
 
-    def max_value(self):
+    def _max_prob_value(self):
         """!
-        \brief maximum factor value for this factor
+        \brief obtain highest yielding domain value and its associated codomain
+        member
+
+        Obtain the highest preference value yielding domain member of this
+        factor with its associated value.
         """
         mx = float("-inf")
         max_val = None
@@ -432,22 +439,158 @@ class Factor(GraphObject):
             if phi_s > mx:
                 mx = phi_s
                 max_val = ss
-        return max_val
+        return max_val, mx
 
-    def partition_value(self, svars):
+    def max_probability(self) -> float:
+        """!
+        \brief maximum preference value for this factor
+
+        \code{.py}
+
+        >>> #
+        >>> Bf = NumCatRVariable(
+        >>>     node_id="B",
+        >>>     input_data={"outcome-values": [10, 50]},
+        >>>     distribution=lambda x: 0.5,
+        >>> )
+        >>> Cf = NumCatRVariable(
+        >>>     node_id="C",
+        >>>     input_data={"outcome-values": [10, 50]},
+        >>>     distribution=lambda x: 0.5,
+        >>> )
+        >>> def phibc(scope_product):
+        >>>     ""
+        >>>     sfs = set(scope_product)
+        >>>     if sfs == set([("B", 10), ("C", 10)]):
+        >>>         return 0.5
+        >>>     elif sfs == set([("B", 10), ("C", 50)]):
+        >>>         return 0.7
+        >>>     elif sfs == set([("B", 50), ("C", 10)]):
+        >>>         return 0.1
+        >>>     elif sfs == set([("B", 50), ("C", 50)]):
+        >>>         return 0.2
+        >>>     else:
+        >>>         raise ValueError("unknown arg")
+
+        >>> bc = Factor(gid="bc", scope_vars=set([Bf, Cf]), factor_fn=phibc)
+        >>> mval = self.bc.max_probability()
+        >>> print(mval)
+        >>> 0.7
+
+        \endcode
+        """
+        mval, mprob = self._max_prob_value()
+        return mprob
+
+    def max_value(self) -> Set[Tuple[str, NumericValue]]:
+        """!
+        \brief maximum factor value for this factor
+
[…]
