### 2021-04-05 D-K-E/graphical-models#5 @D-K-E
All right on it
