### 2021-04-05 D-K-E/graphical-models#5 @D-K-E
I added the `CONTRIBUTING.md` file as requested. How does it look ?
