### 2021-04-05 D-K-E/graphical-models#4 @D-K-E
> sphinx might be ...
I tried my hand at sphinx but couldn't manage to configure it. I am also used to doxygen from other projects, so I would like to stick to what I know for now, if you don't think it bothers the user too much. I find that the generated documentation is a little too complex to navigate for a project of this size for example, but I already have a lot of documented code in doxygen markup as well, so I am hesitating to change it now.

> more information to docstrings: ... method descriptions ...

I am on it. I hope to finish `factor.py` by the end of the week. There are several unused functionality as well. I am hoping to clean them up while I am at it.
