### 2021-04-05 D-K-E/graphical-models#3 @D-K-E
Generally speaking, if some functionality does not have a unittest associated to it, I don't mention it. So far I could not find a good test cases for `MarkovNetwork` and the `ConditionalRandomField`. I am fairly sure of their functionality because all of the work is done by `PGModel` and that is truly tested. I would really appreciate if you can point out to test cases covering markov network and conditional random fields functionality.

I also thought about the real dataset/model suggestion before as well. However `io` and model specification is something I am avoiding for now. I'd like to stick to toy examples for testing the core functionality. I'll think of an io strategy later on.
The problem is since I define factors as functions, it is easier to work with them statically. For a reasonable `io` strategy, I need to generate factor functions dynamically. I might use f string literals or decorators or something else. I have not thought about it yet and I don't plan to at this stage if you don't mind.

What would you prefer to see as a description for toy examples ? Would you mind giving an example ?
