### 2021-04-05 08aefc4
edgetype changed to enum
