### 2021-04-07 #3115 @ankurankan
@D-K-E Thanks for the paper, it was an interesting read but I do have some concerns
which I have listed below.

> Summary: Has a clear description of the high-level functionality and purpose of
> the software for a diverse, non-specialist audience been provided?

From the summary, I am unable to clearly see how the PyGModels package
fits/adds to the current ecosystem of graphical model packages. From my
understanding, the main point being made is that the PyGModels package
implements both the graphical and statistical properties of PGMs. But I think
this true for most of the packages, as graph-theoretical properties can't
really be separated out.  For example, in pgmpy, all the model classes inherit
networkx's graph classes and add properties on top of them. The algorithms also
extensively use the properties of the graphical structure for example computing
the elimination order for variable elimination; order of variables for sampling
algorithms; causal inference works completely on the analysis of graphical
structure; etc.

> A statement of need: Does the paper have a section titled 'Statement of Need'
> that clearly states what problems the software is designed to solve and who the
> target audience is?

From the paper, the two main problems that the package is solving are: 1)
compute posterior, 2) Extensibility. But I think there are a lot of packages
for computing posterior from computationally optimized (eg. pomegranate) to
more readability focused (eg. pgmpy). For extensibility, pgmpy also focuses
quite a bit on that
(https://github.com/pgmpy/pgmpy/blob/dev/examples/Extending%20pgmpy.ipynb).
So, I would have like to see more on how PyGModels adds to these already existing
packages Also from reading this section, it's not clear to me what value for users does
closely following the conventions and definitions add, or who/why would someone
prefer to use this package over the alternatives.

Also, on github's readme, I see that there are some other features implemented
in the package like LWF chain graphs, and graph analysis algorithms, which I
think might be the unique features of this package but there hasn't been much
focus on these in this section.

> State of the field: Do the authors describe how this software compares to other
> commonly-used packages?

I would suggest adding pomegranate (https://github.com/jmschrei/pomegranate) to
this comparison as well as it is one of the most popular packages. I also think
the statement "our implementation ... conforms to the definition provided" is a
bit too strong as all the implementations conform to the standard definitions
and just differ in the data structure being used.

> Quality of writing: Is the paper well written (i.e., it does not require
> editing for structure, language, or writing quality)?

I think the writing needs some work on the content. Reading the current paper,
I feel very little space is currently being used to highlight the use case and
features of the package. I think it would also really help if the paper is
written with a focus on a few main messages that you are trying to convey to
the reader.  Currently, the sections seem to be going in a bit different
directions with not much coherence between them. For example, the summary
mentions that the package is better because of the implementation of the graph-theoretical
 nature of PGMs but this point never comes up again in the rest of
the paper. I think things like easy extensibility, LWF chain, etc should also be
mentioned in the summary as these are the distinguishing features of the
package.
