### 2021-04-07 D-K-E/graphical-models#3 @ankurankan
@D-K-E Thanks for the clarification. For the description, I just meant describing the problem statement that the code example is trying to solve in simple words. An example could be the Monte Hall problem example in pgmpy here: https://github.com/pgmpy/pgmpy/blob/dev/examples/Monty%20Hall%20Problem.ipynb. The notebook gives a context of the Monte Hall problem (which is a toy example) and how you would solve something like that using pgmpy.
