### 2021-04-09 b930935
factor test and documentation added
