+++ gmodels/factor.py
@@ -92,16 +92,23 @@ class Factor(GraphObject):
                 msg += " Negative factors are not allowed"
                 raise ValueError(msg)
 
+        ## random variables belonging to this factor
         self.svars = scope_vars
+
+        ## scope variable hash table
+        self.domain_table = {s.id(): s for s in self.scope_vars()}
+
         if factor_fn is None:
             self.factor_fn = self.marginal_joint
         else:
             self.factor_fn = factor_fn
 
+        ## cartesian product of factor domain
         self.scope_products: List[Set[Tuple[str, NumericValue]]] = list(
             product(*self.vars_domain())
         )
 
+        ## constant normalization value
         self.Z = self.partition_value(self.vars_domain())
 
     def scope_vars(self, f=lambda x: x) -> Set[NumCatRVariable]:
@@ -110,7 +117,7 @@ class Factor(GraphObject):
 
         \param f is a function that transforms the scope of this factor.
 
-        \code
+        \code{.py}
 
         >>> A = NumCatRVariable("A",
         >>>                     input_data={"outcome-values": [True, False]},
@@ -262,10 +269,12 @@ class Factor(GraphObject):
         rvar_filter=lambda x: True,
         value_filter=lambda x: True,
         value_transform=lambda x: x,
-    ):
+    ) -> List[Tuple[Tuple[str, NumericValue]]]:
         """!
-        \brief Compute scope matches for arbitrary domain
-        cartesian product over set of random variables
+        \brief Compute cartesian product over factor domain
+
+        The arguments are there for filtering or manipulating factor domain:
+        \see Factor.fdomain
 
         \code
 
@@ -288,13 +297,16 @@ class Factor(GraphObject):
 
         \endcode
         """
-        svars = cls.fdomain(
-            D=D,
-            rvar_filter=rvar_filter,
-            value_filter=value_filter,
-            value_transform=value_transform,
+        return list(
+            product(
+                *cls.fdomain(
+                    D=D,
+                    rvar_filter=rvar_filter,
+                    value_filter=value_filter,
+                    value_transform=value_transform,
+                )
+            )
         )
-        return list(product(*svars))
 
     def factor_domain(
         self,
@@ -304,6 +316,7 @@ class Factor(GraphObject):
     ) -> list:
         """!
         \brief \see Factor.matches(rvar_filter, value_filter, value_transform)
+
         For a factor phi(A,B) return factor function's domain values, such as:
 
         \f[phi(A,B)\f]
@@ -317,7 +330,12 @@ class Factor(GraphObject):
 
         >>> set(("A", a1), ("B", b2)),...
         """
-        return list(product(*self.vars_domain()))
+        return self.matches(
+            D=self.scope_vars(),
+            rvar_filter=rvar_filter,
+            value_filter=value_filter,
+            value_transform=value_transform,
+        )
 
     def vars_domain(
         self,
@@ -346,22 +364,64 @@ class Factor(GraphObject):
         Each value in domain comes with an identifier of its random variable.
         From these identifiers, we obtain set of random variables attested in
         factor domain.
+
+        \param domain list of arbitrary domain values
+        \exception ValueError We raise value error when the argument domain
+        value array is not a subset of the domain of the factor, since we have
+        no way of obtaining random variable that is not inside the scope of
+        this factor.
+
+        \return set of random variables implied by the given list of domain
+        values
+
+        \code{.py}
+
+        >>> Af = NumCatRVariable(
+        >>>    node_id="A",
+        >>>    input_data={"outcome-values": [10, 50]},
+        >>>    distribution=lambda x: 0.5,
+        >>> )
+        >>> Bf = NumCatRVariable(
+        >>>    node_id="B",
+        >>>    input_data={"outcome-values": [10, 50]},
+        >>>    distribution=lambda x: 0.5,
[…]
