### 2021-04-10 edc75c5
factor documentation more or less done
