+++ gmodels/factor.py
@@ -7,7 +7,7 @@ Defining a factor from Koller and Friedman 2009, p. 106-107
 from gmodels.gtypes.graphobj import GraphObject
 from gmodels.randomvariable import NumCatRVariable, NumericValue
 
-from typing import Set, Callable, Optional, List, Union, Tuple
+from typing import Set, Callable, Optional, List, Union, Tuple, FrozenSet
 from itertools import product, combinations
 from uuid import uuid4
 from pprint import pprint
@@ -223,7 +223,7 @@ class Factor(GraphObject):
         rvar_filter=lambda x: True,
         value_filter=lambda x: True,
         value_transform=lambda x: x,
-    ) -> List[Set[Tuple[str, NumericValue]]]:
+    ) -> List[FrozenSet[Tuple[str, NumericValue]]]:
         """!
         \brief Get factor domain Val(D) D being a set of random variables
 
@@ -249,8 +249,8 @@ class Factor(GraphObject):
         >>> fmatches = Factor.fdomain(D=D)
         >>> print(fmatches)
 
-        >>> [set(("A", True), ("A", True)),
-        >>>  set(("B", True), ("B", False)),
+        >>> [frozenset(("A", True), ("A", True)),
+        >>>  frozenset(("B", True), ("B", False)),
         >>> ]
 
         \endcode
@@ -259,7 +259,7 @@ class Factor(GraphObject):
         return [
             s.value_set(value_filter=value_filter, value_transform=value_transform)
             for s in D
-            if rvar_filter(s) is True
+            if rvar_filter(s)
         ]
 
     @classmethod
@@ -269,14 +269,14 @@ class Factor(GraphObject):
         rvar_filter=lambda x: True,
         value_filter=lambda x: True,
         value_transform=lambda x: x,
-    ) -> List[Tuple[Tuple[str, NumericValue]]]:
+    ) -> List[FrozenSet[Tuple[str, NumericValue]]]:
         """!
         \brief Compute cartesian product over factor domain
 
         The arguments are there for filtering or manipulating factor domain:
         \see Factor.fdomain
 
-        \code
+        \code{.py}
 
         >>> A = NumCatRVariable("A",
         >>>                     input_data={"outcome-values": [True, False]},
@@ -291,29 +291,27 @@ class Factor(GraphObject):
         >>> fmatches = Factor.matches(D=D)
         >>> print(fmatches)
 
-        >>> [(("A", True), ("B", True)),
-        >>>  (("A", True), ("B", False)),
-        >>>  (("A", False), ("B", True)), (("A", False), ("B", False))]
+        >>> [frozenset([("A", True), ("B", True)]),
+        >>>  frozenset([("A", True), ("B", False)]),
+        >>>  frozenset([("A", False), ("B", True)]), 
+        >>>  frozenset([("A", False), ("B", False)])]
 
         \endcode
         """
-        return list(
-            product(
-                *cls.fdomain(
-                    D=D,
-                    rvar_filter=rvar_filter,
-                    value_filter=value_filter,
-                    value_transform=value_transform,
-                )
-            )
+        domain_values = cls.fdomain(
+            D=D,
+            rvar_filter=rvar_filter,
+            value_filter=value_filter,
+            value_transform=value_transform,
         )
+        return [frozenset(s) for s in list(product(*domain_values))]
 
     def factor_domain(
         self,
         rvar_filter=lambda x: True,
         value_filter=lambda x: True,
         value_transform=lambda x: x,
-    ) -> list:
+    ) -> List[FrozenSet[Tuple[str, NumericValue]]]:
         """!
[…]
