+++ paper/paper.md
@@ -26,7 +26,8 @@ libraries implement PGMs in a way that ignores their graphical nature.
 nature of PGMs, thereby giving `PyGModels`' instantiated objects both
 graph-theoretical and statistical properties, which allows users to explore
 and test inference algorithms that are rooted in both graph theory or
-statistics.
+statistics. `PyGModels` also implements several algorithms of interest on a
+LWF chain graphs, also known as mixed graphs.
 
 # Statement of Need
 
@@ -40,14 +41,24 @@ Let us try to demonstrate the need for `PyGModels` by a use case.
 One has a set of categorical random variables in the form of a function
 specified by a probability distribution. One has a set of edges that encode a
 certain independence assumption over her random variables, and one has a set
-of factors, factorizes a certain probability distribution over her entire
+of factors, that factorizes a certain probability distribution over her entire
 graph. Given these, `PyGModels` might solve two major issues for the
 researcher or the student alike:
 
-- Compute posterior probability distribution or most probable explanation
+1. Compute posterior probability distribution or most probable explanation
   given certain evidence.
 
-- Provide a basis for creating new algorithms of inference.
+2. Provide a basis for creating new algorithms of inference.
+
+If the independence assumptions over the random variables requires the graph
+to be a LWF chain graph where the graph can have both directed and undirected
+edges, `PyGModels` can also solve:
+
+3. Decomposing the chain graph into chain components
+
+4. Moralizing the chain graph into a Markov Network.
+
+5. Decomposing the chain graph into Conditional Random Fields.
 
 Though the first issue is not irrelevant, the forte of `PyGModels` is the
 second issue due to its lightweight nature and its direct embodiment of
@@ -57,14 +68,14 @@ graph theoretic (we follow mostly Diestel [see @Diestel_2017] for graph
 theoretic conventions and definitions; most of the graph algorithms come
 from K. Erciyes [see @Erciyes_2018] and S. Even [see @Even_Guy_Even_2012];
 exact pages are cited in doc strings of related functions inside the
-source code) considerations in the same base class. The entire library
-depends only on python standard library which makes it very extendible and
-easy to integrate and adapt to other projects as well. Through its
-rigorous adoption of mathematical definitions of involved concepts, it
-becomes feasible to extend arbitrary factors through their point wise
-product, or apply common graph analysis algorithms such as finding
-articulation points or bridges, or finding an optimal path defined by a
-cost function.
+source code) considerations in the same base class. 
+
+The entire library depends only on python standard library which makes it very
+extendible and easy to integrate and adapt to other projects as well. Through
+its rigorous adoption of mathematical definitions of involved concepts, it
+becomes feasible to extend arbitrary factors through their point wise product,
+or apply common graph analysis algorithms such as finding articulation points
+or bridges, or finding an optimal path defined by a cost function.
 
 # Applications and Similar Works
 
@@ -84,12 +95,13 @@ Other open sourced python libraries about PGMs include the following:
 
 - `pyfac` [see @Lester_2016]
 
-The most popular and goto library is `pgmpy`. It has also been used in several
-publications [see @Ankan_Panda_2015a; also @Ankan_Panda_2015b]. Its
-functionality is covered with a nice test suite as well. Overall it is a
-reliable library for using PGMs in production. It comes with several
-dependencies that can be heavy handed (`pytorch` for example) for exploratory
-use though.
+- `pomegranate` [see @Schreiber_2018]
+
+The most popular and goto libraries are `pgmpy` and `pomegranate`. Both of
+them have also been used in several publications [see @Ankan_Panda_2015a; also
+@Ankan_Panda_2015b; and @Schreiber_2018]. Their functionalities are covered
[…]
