### 2021-04-10 927b087
changed pgmpy to package
