### 2021-04-10 3b73329
changed math formulas
