### 2021-04-10 #3115 @D-K-E
@ankurankan Thanks for the detailed review. I pushed some changes to the paper according to your concerns.

Here is a list of changes in accordance with your comments:

---

## Summary

I added the phrase `PyGModels also implements several algorithms of interest on
LWF chain graphs, also known as mixed graphs.` to the summary hoping that it would underline the main contribution of this package to the ecosystem without introducing too much technical jargon.

>From the summary ... fits/adds to the current ecosystem...

The main problem here is that since the current packages like `pomegranate` and `pgmpy` are quite well developed and production ready, I am not sure if it is feasible to underline the contribution of `PyGModels`, inference on chain graphs which covers relatively a marginal use case outside of PGM community,  to a non specialist audience without getting into technical details. The limitation on LWF chain graphs comes from `NetworkX` since it does not yet support mixed graphs as stated in the [issue](https://github.com/networkx/networkx/issues/1168). However, since chain graphs can be decomposed to conditional random fields, the inference on chain graphs can be supported via CRFs. Hence the willing user might find ways of doing inference on chain graphs using existing packages as well. Yet again, I think this level of discussion is too technical for a non specialist audience but let's ask @dfm whether I should incorporate all this into the paper.

## Statement of Need

As per suggested I increased the number of issues that `PyGModels` solve. I added
```
3. Decomposing the chain graph into chain components
4. Moralizing the chain graph into a Markov Network.
5. Decomposing the chain graph into Conditional Random Fields.
```

>...what value for users does closely following the conventions and definitions add...

Bluntly speaking not a lot in practical sense, however it might decrease the friction between text books and code. Hence it has a pedagogical value. The documentation is not very fulfilling yet to truly accomplish this though.

>adding pomegranate ... is a bit too strong...

I added `pomegranate` to the comparison, and added the phrase:
```
The last aspect is also the case for other packages, however PyGModels differs from them with respect to the data structure used in the implementation.
```

---

>I feel very little space is currently being used to highlight the use case and
features of the package

As stated, I mentioned LWF chain graphs in the summary as you've requested. Since the recommended word count is between 250 - 1000 for a paper according to journal's [documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and the paper's count is 1037, I don't think it would be wise for me to add new sections. I can remove certain parts of the paper, if you think that's good choice. If you think removal is the way to go, would mind suggesting some parts ?
