### 2021-04-17 17c5109
new tests added. Graph documentation started
