+++ gmodels/gtypes/graph.py
@@ -18,7 +18,7 @@ import math
 class Graph(GraphObject):
     """!
     Simple finite graph
-    \f[ G = (V, E) \f] where \f[ V \f] is the vertex set and \f[ E \f] is the edge set.
+    \f$ G = (V, E) \f$ where \f$ V \f$ is the vertex set and \f$ E \f$ is the edge set.
     """
 
     def __init__(
@@ -35,10 +35,34 @@ class Graph(GraphObject):
         \param nodes a node/vertex set.
         \param edges an edge set.
 
+        \throws ValueError If the graph is trivial, we raise a value error, as
+        most algorithms don't work with trivial graphs.
+
         We construct the graph from given node and edge set. For quick look up
         we store them in hash tables. The gdata member also stores an edge list
         representation, in order to facilitate some of the basic look up
         functionality concerning neighbours of vertices.
+
+        \code{.py}
+
+        >>> a = Node("a", {})  # b
+        >>> b = Node("b", {})  # c
+        >>> f = Node("f", {})  # d
+        >>> e = Node("e", {})  # e
+        >>> ab = Edge(
+        >>>    "ab", start_node=a, end_node=b, edge_type=EdgeType.UNDIRECTED
+        >>> )
+        >>> be = Edge(
+        >>>    "be", start_node=b, end_node=e, edge_type=EdgeType.UNDIRECTED
+        >>> )
+        >>> graph = Graph(
+        >>>     "graph",
+        >>>     data={"my": "graph", "data": "is", "very": "awesome"},
+        >>>     nodes=set([a, b, e, f]),
+        >>>     edges=set([ab, be]),
+        >>> )
+
+        \endcode
         """
         super().__init__(oid=gid, odata=data)
         self._nodes: Optional[Dict[str, Node]] = None
@@ -73,6 +97,8 @@ class Graph(GraphObject):
         \brief We construct the graph from given edge set using a random id.
 
         See \see Graph for more information
+
+        We obtain nodes from edges then pass them to graph constructor.
         """
         nodes: Set[Node] = set()
         for e in edges:
@@ -85,6 +111,12 @@ class Graph(GraphObject):
         """!
         \brief We construct the graph from given node, and edge sets using a random id.
         \see Graph for more information
+
+        \param edges set of edges
+        \param nodes set of nodes
+
+        We iterate over set of edges and add the nodes that are not inside the
+        node set. We pass both parameters to Graph constructor afterwards.
         """
         nodes = set(nodes)
         for e in edges:
@@ -100,7 +132,7 @@ class Graph(GraphObject):
         \param es set of edges
 
         We assume that node set and edge set might contain different nodes,
-        that is \f[ V[G] = V[ns] \cup V[es] \f]
+        that is \f$ V[G] = V[ns] \cup V[es] \f$
         We combine nodes given in both sets to create a final set of nodes
         for the graph
         """
@@ -118,9 +150,61 @@ class Graph(GraphObject):
         #
         self._nodes = {n.id(): n for n in nodes}
 
-    def to_adjmat(self, vtype=int):
+    def to_adjmat(self, vtype=int) -> Dict[Tuple[str, str], int]:
         """!
         \brief Transform adjacency list to adjacency matrix representation
+
[…]
