### 2021-04-18 ed194df
added some docs to random variable
