### 2021-04-19 50c443d
improvement in random variable documentation
