+++ gmodels/randomvariable.py
@@ -8,6 +8,19 @@ import math
 from uuid import uuid4
 from random import choice
 
+Outcome = Any
+Value = Any
+NumericValue = float
+
+
+class PossibleOutcomes:
+    """!
+    \brief set of possible outcomes from Koller, Friedman 2009, p. 15, 20
+    """
+
+    def __init__(self, omega: FrozenSet[Outcome]):
+        self.data = omega
+
 
 class RandomVariable(Node):
     """!
@@ -24,7 +37,7 @@ class RandomVariable(Node):
     """
 
     def __init__(
-        self, node_id: str, data: Any, f: Callable[[Any], Any] = lambda x: x,
+        self, node_id: str, data: Any, f: Callable[[Outcome], Value] = lambda x: x,
     ):
         """!
         \brief Constructor of a random variable
@@ -40,11 +53,6 @@ class RandomVariable(Node):
         raise NotImplementedError
 
 
-Outcome = Any
-Value = Any
-NumericValue = float
-
-
 class CatRandomVariable(RandomVariable):
     """!
     a discrete random variable
@@ -53,17 +61,57 @@ class CatRandomVariable(RandomVariable):
     def __init__(
         self,
         node_id: str,
-        input_data: Dict[str, Outcome],
+        input_data: Dict[str, Any],
         f: Callable[[Outcome], Value] = lambda x: x,
         distribution: Callable[[Value], float] = lambda x: 1.0,
     ):
-        ""
+        """!
+        \brief Constructor for categorical/discrete random variable
+
+        \param distribution a function that takes in a value from codomain of
+        the random variable and outputs a value in the range [0,1].
+
+        \throws ValueError We raise a value error if the probability values
+        associated to outcomes add up to a value bigger than one.
+
+        For other parameters and the definition of a random variable \see
+        RandomVariable .
+
+        A simple data specification is provided for passing evidences and
+        input.
+        The possible outcomes key holds a set of values belonging to space of
+        possible outcomes. If the input data just contains a key as
+        'possible-outcomes', we suppose that it contains a PossibleOutcomes
+        object which represents the space of all possible outcomes of the
+        measurable event set associated to random variable.
+
+        The function associated to our random variable transforms the set of
+        possible outcomes to values as per its definition in Koller, Friedman,
+        2009, p. 20. Lastly we check whether obtained, or associated
+        outcome-values satisfy the probability rule by checking if the
+        probabilities associated to these values add up to one.
+
+        \code{.py}
+
+        >>> students = PossibleOutcomes(frozenset(["student_1", "student_2"]))
+        >>> grade_f = lambda x: "F" if x == "student_1" else "A"
+        >>> grade_distribution = lambda x: 0.1 if x == "F" else 0.9
+        >>> indata = {"possible-outcomes": students}
+        >>> rvar = CatRandomVariable(
+        >>>    input_data=indata, 
+        >>>    node_id="myrandomvar", 
+        >>>    f=grade_f,
+        >>>    distribution=grade_distribution
+        >>> )
+
+        \endcode
+        """
         data = {}
         data.update(input_data)
-        if "outcomes" in input_data:
-            data["outcome-values"] = {
-                (i, v): f(v) for i, v in enumerate(input_data["outcomes"])
-            }
+        if "possible-outcomes" in input_data:
+            data["outcome-values"] = frozenset(
+                [f(v) for v in input_data["possible-outcomes"].data]
+            )
         super().__init__(node_id=node_id, data=data, f=f)
         if "outcome-values" in data:
             psum = sum(list(map(distribution, data["outcome-values"])))
@@ -71,10 +119,14 @@ class CatRandomVariable(RandomVariable):
                 raise ValueError("probability sum bigger than 1 or smaller than 0")
[…]
