+++ README.md
@@ -93,7 +93,6 @@ Ran 179 tests in 0.666s
 OK
 ```
 
-
 ## Guide for Contributors
 
 See [Contributing.md](CONTRIBUTING.md)
