+++ docs/mainp.md
@@ -5,7 +5,7 @@
 
 [![DOI](https://zenodo.org/badge/321839625.svg)](https://zenodo.org/badge/latestdoi/321839625)
 
-See doxygen generated [documentation](https://viva-lambda.github.io/graphical-models/)
+See doxygen generated [documentation](https://d-k-e.github.io/graphical-models/)
 
 
 The source code of this library aims to be accessible to all those interested
@@ -95,244 +95,21 @@ OK
 
 ## Guide for Contributors
 
-As of now, most of the functions are unit tested. The test suit contains
-around 170 tests covering most of the important functionality. However, there
-can never be too much tests, so feel free to create a pull request with some
-of your own.
-
-Another area of improvement is documentation. As of now, we lack usage
-examples for graph theoretical functionality of the library. Some functions
-can also use more elaborate docstrings. Notice that we are not using sphinx or
-other regular pythonic documentation generators. We are using
-[doxygen](https://www.doxygen.nl/index.html). For the willing user, we provide
-two templates for filling out docstrings in case she is not familiar with
-doxygen way of doing things:
-
-- Partial template:
-
-```python
-
-def my_function(myarg1: str, myarg2: int) -> str:
-    """!
-    \brief One line explanation of functionality
-
-    \param myarg1 description of the argument
-    \param myarg2 description of the argument
-
-    \return description of the returned value
-    """
-    return myarg1 + str(myarg2)
-```
-
-- Full template:
-
-\verbatim
-
-def my_function(myarg1: str, myarg2: int) -> str:
-    """!
-    \brief One line explanation of functionality
-
-    Long multilined
-    description of
-    functionality
-
-    \param myarg1 description of the argument
-    \param myarg2 description of the argument
-
-    \exception TypeError description of the exception 
-
-    \return description of the returned value
-
-    \code{.py}
-
-    >>> a = my_function("Lucky number is ", 7)
-    >>> print(a)
-    >>> "Lucky number is 7"
-
-    \endcode
-
-    """
-
-    if isinstance(myarg2, int) is False:
-        raise TypeError(
-            "myarg2 " + str(myarg2) + " is of type " +
-            str(type(myarg2))
-            )
-    return myarg1 + str(myarg2)
-
-\endverbatim
-
-Besides adding a documentation, you can also add other inference strategies.
-
-Just file an issue in the case of doubt or signal your intent and we can
-discuss the rest.
-
-
-
-## Currently Supported Models
-
-There are currently 4 primary models that might interest the general public:
-
-### PGModel
-
-`PGModel`, a graph with defined as `G=(V,E,F)` where `V` is vertex set
-composed of `NumCatRVariable`s which are numeric categorical random variables,
-`E` is edge set, and `F` is factor set.  
-If `F` is none the factor set is deduced from edge set where each edge is
-considered as a factor whose scope is the set of incident nodes to edge.
-Sum-Product and Max-Product variable elimination are supported by `PGModel`.
-Evidence encoded in nodes are directly used for reducing factors during the
-instantiation of `PGModel`.
-Additional evidence can be provided at query time as well for conditional and
-mpe queries. `Marginal Map` queries are not yet supported.
-
-Usage:
-
-\code{.py}
-
-from gmodels.pgmodel import PGModel
-from gmodels.gtypes.edge import Edge, EdgeType
[…]
