### 2021-04-22 78ff981
new random variable tests and documentation to random var
