### 2021-04-22 5f42562
new random variable tests and documentation to random var
