### 2021-04-22 2d550d8
added new documentation to random variable and new tests
