+++ gmodels/randomvariable.py
@@ -245,6 +245,10 @@ class CatRandomVariable(RandomVariable):
 class NumCatRVariable(CatRandomVariable):
     """!
     \brief Numerical categorical random variable object
+
+    This is mostly the same as CatRandomVariable. The main difference is that
+    the function associated to random variable produces a numeric value as an
+    outcome
     """
 
     def __init__(
@@ -263,13 +267,37 @@ class NumCatRVariable(CatRandomVariable):
         is numeric, that is it can be integer or float. For facilitating
         operations we treat everything as float.
 
+        \code{.py}
+        >>> nid1 = "rvar1"
+        >>> input_data = {
+        >>>    "intelligence": {"outcome-values": [0.1, 0.9], "evidence": 0.9},
+        >>>    "grade": {"outcome-values": [0.2, 0.4, 0.6], "evidence": 0.2},
+        >>>    "dice": {"outcome-values": [i for i in range(1, 7)], "evidence": 1.0 / 6},
+        >>> }
+
+        >>> def intelligence_dist(intelligence_value: float) -> float:
+        >>>    if intelligence_value == 0.1:
+        >>>        return 0.7
+        >>>    elif intelligence_value == 0.9:
+        >>>        return 0.3
+        >>>    else:
+        >>>        return 0.0
+
+        >>> # intelligence
+        >>> intelligence = NumCatRVariable(
+        >>>    node_id=nid1,
+        >>>    input_data=input_data["intelligence"],
+        >>>    distribution=intelligence_dist,
+        >>> )
+        \endcode
+
         """
         super().__init__(
             node_id=node_id, input_data=input_data, f=f, distribution=distribution
         )
 
     @staticmethod
-    def type_check(other: Any) -> bool:
+    def type_check(other: Any) -> None:
         """!
         \brief simple function for checking whether the other is also a
         NumCatRVariable
@@ -283,49 +311,227 @@ class NumCatRVariable(CatRandomVariable):
             raise TypeError(
                 "other arg must be of type NumCatRVariable, it is " + type(other)
             )
-        return True
 
-    def evidence_check(self) -> bool:
+    def has_evidence(self) -> None:
         """!
         \brief Check if any evidence is associated with this random variable
+
+        \throws ValueError We raise a value error if there is no evidence
+        associated to random variable.
         """
         data = self.data()
         if "evidence" not in data:
             msg = "Evidence " + " could not be found with in"
             msg += " attributed data of this random variable"
             raise ValueError(msg)
-        return True
 
-    def max(self):
+    def max(self) -> float:
         """!
         \brief maximum marginal value
+
+        We return the highest marginal/probability.
+
+        \code{.py}
+        >>> nid1 = "rvar1"
+        >>> input_data = {
+        >>>    "intelligence": {"outcome-values": [0.1, 0.9], "evidence": 0.9},
+        >>>    "grade": {"outcome-values": [0.2, 0.4, 0.6], "evidence": 0.2},
+        >>>    "dice": {"outcome-values": [i for i in range(1, 7)], "evidence": 1.0 / 6},
+        >>> }
+
+        >>> def intelligence_dist(intelligence_value: float) -> float:
+        >>>    if intelligence_value == 0.1:
+        >>>        return 0.7
+        >>>    elif intelligence_value == 0.9:
+        >>>        return 0.3
+        >>>    else:
+        >>>        raise ValueError(
+        >>>            "intelligence_value does not belong to possible outcomes"
+        >>>        )
+
+        >>> # intelligence
+        >>> intelligence = NumCatRVariable(
[…]
