### 2021-04-29 41b2c0e
added markov network tests
