+++ gmodels/lwfchain.py
@@ -284,3 +284,342 @@ class LWFChainGraph(PGModel):
         """!
         """
         return self.is_parent_of(n1=parent, n2=child)
+
+
+"""!
+\page lwfchaingraphexample LWF Chain Graph Usage
+
+LWF Chain Graph, also known as Partially Directed Model, is a probabilistic
+graphical model. It's distinctive nature is best explained by Koller, Friedman
+2009, p. 148:
+<blockquote>
+An edge between two nodes in the same chain component must be undirected, while
+an edge between two nodes in different chain components must be directed.
+</blockquote>
+
+S. Lauritzen, one of the main contributors (the L in LWF) of the subject, had
+written extensively on the subject. However probably the most through treatment
+of the subject is Lauritzen 1996, p. 158 - 220. He had also provided the causal
+interpretation of LWF chain graphs in [Lauritzen 2002](http://doi.wiley.com/10.1111/1467-9868.00340).
+The standard inference strategies on chain graphs are best explained in Cowell
+2005 and more recently in Dechter 2019.
+
+Usage:
+
+\code{.py}
+
+idata = {"outcome-values": [True, False]}
+A = NumCatRVariable(
+       node_id="A", input_data=idata, distribution=lambda x: 0.5
+)
+B = NumCatRVariable(
+       node_id="B", input_data=idata, distribution=lambda x: 0.5
+)
+C = NumCatRVariable(
+       node_id="C", input_data=idata, distribution=lambda x: 0.5
+)
+D = NumCatRVariable(
+       node_id="D", input_data=idata, distribution=lambda x: 0.5
+)
+E = NumCatRVariable(
+       node_id="E", input_data=idata, distribution=lambda x: 0.5
+)
+F = NumCatRVariable(
+       node_id="F", input_data=idata, distribution=lambda x: 0.5
+)
+G = NumCatRVariable(
+       node_id="G", input_data=idata, distribution=lambda x: 0.5
+)
+H = NumCatRVariable(
+       node_id="H", input_data=idata, distribution=lambda x: 0.5
+)
+I = NumCatRVariable(
+       node_id="I", input_data=idata, distribution=lambda x: 0.5
+)
+K = NumCatRVariable(
+       node_id="K", input_data=idata, distribution=lambda x: 0.5
+)
+L = NumCatRVariable(
+       node_id="L", input_data=idata, distribution=lambda x: 0.5
+)
+#
+#  Cowell 2005, p. 110
+#
+#   A                      E---+
+#   |                          |
+#   +----+                 F <-+
+#        |                 |
+#   B <--+---> C --> D <---+
+#   |                |
+#   +---> H <--------+----> G
+#   |     |
+#   +---> I
+#
+AB_c = Edge(
+  edge_id="AB",
+  start_node=A,
+  end_node=B,
+  edge_type=EdgeType.DIRECTED,
+)
+AC_c = Edge(
+  edge_id="AC",
+  start_node=A,
+  end_node=C,
+  edge_type=EdgeType.DIRECTED,
+)
+CD_c = Edge(
+  edge_id="CD",
+  start_node=C,
+  end_node=D,
+  edge_type=EdgeType.DIRECTED,
+)
+EF_c = Edge(
+  edge_id="EF",
+  start_node=E,
+  end_node=F,
+  edge_type=EdgeType.DIRECTED,
+)
+FD_c = Edge(
+  edge_id="FD",
+  start_node=F,
+  end_node=D,
+  edge_type=EdgeType.DIRECTED,
+)
+DG_c = Edge(
+  edge_id="DG",
+  start_node=D,
+  end_node=G,
+  edge_type=EdgeType.DIRECTED,
+)
[…]
