### 2021-04-29 fc9ef6b
added markov network tests
