+++ test/test_markov.py
@@ -1,0 +1,220 @@
+"""!
+Test markov network
+"""
+
+from gmodels.randomvariable import NumCatRVariable
+from gmodels.markov import MarkovNetwork, ConditionalRandomField
+from gmodels.gtypes.edge import Edge, EdgeType
+from gmodels.gtypes.undigraph import UndiGraph
+from gmodels.factor import Factor
+from uuid import uuid4
+import pdb
+
+import unittest
+
+
+class MarkovTest(unittest.TestCase):
+    ""
+
+    def setUp(self):
+        ""
+        idata = {
+            "A": {"outcome-values": [True, False]},
+            "B": {"outcome-values": [True, False]},
+            "C": {"outcome-values": [True, False]},
+            "D": {"outcome-values": [True, False]},
+        }
+
+        # misconception example
+        self.A = NumCatRVariable(
+            node_id="A", input_data=idata["A"], distribution=lambda x: 0.5
+        )
+        self.B = NumCatRVariable(
+            node_id="B", input_data=idata["B"], distribution=lambda x: 0.5
+        )
+        self.C = NumCatRVariable(
+            node_id="C", input_data=idata["C"], distribution=lambda x: 0.5
+        )
+        self.D = NumCatRVariable(
+            node_id="D", input_data=idata["D"], distribution=lambda x: 0.5
+        )
+        self.AB = Edge(
+            edge_id="AB",
+            edge_type=EdgeType.UNDIRECTED,
+            start_node=self.A,
+            end_node=self.B,
+        )
+        self.AD = Edge(
+            edge_id="AD",
+            edge_type=EdgeType.UNDIRECTED,
+            start_node=self.A,
+            end_node=self.D,
+        )
+        self.DC = Edge(
+            edge_id="DC",
+            edge_type=EdgeType.UNDIRECTED,
+            start_node=self.D,
+            end_node=self.C,
+        )
+        self.BC = Edge(
+            edge_id="BC",
+            edge_type=EdgeType.UNDIRECTED,
+            start_node=self.B,
+            end_node=self.C,
+        )
+
+        def phi_AB(scope_product):
+            ""
+            ss = frozenset(scope_product)
+            if ss == frozenset([("A", False), ("B", False)]):
+                return 30.0
+            elif ss == frozenset([("A", False), ("B", True)]):
+                return 5.0
+            elif ss == frozenset([("A", True), ("B", False)]):
+                return 1.0
+            elif ss == frozenset([("A", True), ("B", True)]):
+                return 10.0
+            else:
+                raise ValueError("product error")
+
+        def phi_BC(scope_product):
+            ""
+            ss = frozenset(scope_product)
+            if ss == frozenset([("B", False), ("C", False)]):
+                return 100.0
+            elif ss == frozenset([("B", False), ("C", True)]):
+                return 1.0
+            elif ss == frozenset([("B", True), ("C", False)]):
+                return 1.0
+            elif ss == frozenset([("B", True), ("C", True)]):
+                return 100.0
+            else:
+                raise ValueError("product error")
+
+        def phi_CD(scope_product):
+            ""
+            ss = frozenset(scope_product)
+            if ss == frozenset([("C", False), ("D", False)]):
+                return 1.0
+            elif ss == frozenset([("C", False), ("D", True)]):
+                return 100.0
+            elif ss == frozenset([("C", True), ("D", False)]):
+                return 100.0
+            elif ss == frozenset([("C", True), ("D", True)]):
+                return 1.0
+            else:
[…]
