+++ gmodels/pgmodel.py
@@ -386,105 +386,3 @@ class PGModel(Graph):
             max_assign = diff.pop()
             max_assignments[max_assign[0]] = max_assign[1]
         return max_assignments
-
-
-"""!
-\page pgmodelusage PGModel Usage
-
-
-`PGModel`, a graph with defined as `G=(V,E,F)` where `V` is vertex set composed
-of `NumCatRVariable`s which are numeric categorical random variables, `E` is
-edge set, and `F` is factor set. If `F` is none the factor set is deduced from
-edge set where each edge is considered as a factor whose scope is the set of
-incident nodes to edge. Sum-Product and Max-Product variable elimination are
-supported by `PGModel`. Evidence encoded in nodes are directly used for
-reducing factors during the instantiation of `PGModel`. Additional evidence can
-be provided at query time as well for conditional and most probable explanation
-queries. `Marginal Map` queries are not yet supported.
-
-Usage:
-
-\code{.py}
-
-from gmodels.pgmodel import PGModel
-from gmodels.gtypes.edge import Edge, EdgeType
-from gmodels.factor import Factor
-from gmodels.randomvariable import NumCatRVariable
-# Example adapted from Darwiche 2009, p. 140
-idata = {
-            "a": {"outcome-values": [True, False]},
-            "b": {"outcome-values": [True, False]},
-            "c": {"outcome-values": [True, False]},
-        }
-a = NumCatRVariable(
-            node_id="a", input_data=idata["a"], distribution=lambda x: 0.6 if x else 0.4
-        )
-b = NumCatRVariable(
-    node_id="b", input_data=idata["b"], distribution=lambda x: 0.5 if x else 0.5
-)
-c = NumCatRVariable(
-    node_id="c", input_data=idata["c"], distribution=lambda x: 0.5 if x else 0.5
-)
-ab = Edge(
-    edge_id="ab",
-    edge_type=EdgeType.UNDIRECTED,
-    start_node=a,
-    end_node=b,
-)
-bc = Edge(
-    edge_id="bc",
-    edge_type=EdgeType.UNDIRECTED,
-    start_node=b,
-    end_node=c,
-)
-def phi_ba(scope_product):
-    ""
-    ss = set(scope_product)
-    if ss == set([("a", True), ("b", True)]):
-        return 0.9
-    elif ss == set([("a", True), ("b", False)]):
-        return 0.1
-    elif ss == set([("a", False), ("b", True)]):
-        return 0.2
-    elif ss == set([("a", False), ("b", False)]):
-        return 0.8
-    else:
-        raise ValueError("product error")
-def phi_cb(scope_product):
-    ""
-    ss = set(scope_product)
-    if ss == set([("c", True), ("b", True)]):
-        return 0.3
-    elif ss == set([("c", True), ("b", False)]):
-        return 0.5
-    elif ss == set([("c", False), ("b", True)]):
-        return 0.7
-    elif ss == set([("c", False), ("b", False)]):
-        return 0.5
-    else:
-        raise ValueError("product error")
-def phi_a(scope_product):
-    s = set(scope_product)
-    if s == set([("a", True)]):
-        return 0.6
-    elif s == set([("a", False)]):
-        return 0.4
-    else:
-        raise ValueError("product error")
-ba_f = Factor(gid="ba", scope_vars=set([b, a]), factor_fn=phi_ba)
-cb_f = Factor(gid="cb", scope_vars=set([c, b]), factor_fn=phi_cb)
-a_f = Factor(gid="a", scope_vars=set([a]), factor_fn=phi_a)
-pgm = PGModel(
-    gid="pgm",
-    nodes=set([a, b, c]),
[…]
