+++ gmodels/markov.py
@@ -314,144 +314,3 @@ class ConditionalRandomField(MarkovNetwork):
             return p_yx_i / z_i
 
         return Factor(gid=str(uuid4()), factor_fn=phi_cond, scope_vars=self.X)
-
-
-"""!
-\page markovnetworkexample Markov Network Usage
-
-Markov Network is defined by Koller, Friedman 2009, p. 103 is
-<blockquote>
-As in a Bayesian network, the nodes in the graph of a Markov network represent
-the variables, and the edges correspond to a notion of direct probabilistic
-interaction between the neighboring variables — an interaction that is not
-mediated by any other variable in the network.
-</blockquote>
-
-Usage:
-
-\code{.py}
-
-idata = {
-    "A": {"outcome-values": [True, False]},
-    "B": {"outcome-values": [True, False]},
-    "C": {"outcome-values": [True, False]},
-    "D": {"outcome-values": [True, False]},
-}
-                                                                   
-# misconception example
-A = NumCatRVariable(
-    node_id="A", input_data=idata["A"], distribution=lambda x: 0.5
-)
-B = NumCatRVariable(
-    node_id="B", input_data=idata["B"], distribution=lambda x: 0.5
-)
-C = NumCatRVariable(
-    node_id="C", input_data=idata["C"], distribution=lambda x: 0.5
-)
-D = NumCatRVariable(
-    node_id="D", input_data=idata["D"], distribution=lambda x: 0.5
-)
-AB = Edge(
-    edge_id="AB",
-    edge_type=EdgeType.UNDIRECTED,
-    start_node=A,
-    end_node=B,
-)
-AD = Edge(
-    edge_id="AD",
-    edge_type=EdgeType.UNDIRECTED,
-    start_node=A,
-    end_node=D,
-)
-DC = Edge(
-    edge_id="DC",
-    edge_type=EdgeType.UNDIRECTED,
-    start_node=D,
-    end_node=C,
-)
-BC = Edge(
-    edge_id="BC",
-    edge_type=EdgeType.UNDIRECTED,
-    start_node=B,
-    end_node=C,
-)
-                                                                   
-def phi_AB(scope_product):
-    ""
-    ss = frozenset(scope_product)
-    if ss == frozenset([("A", False), ("B", False)]):
-        return 30.0
-    elif ss == frozenset([("A", False), ("B", True)]):
-        return 5.0
-    elif ss == frozenset([("A", True), ("B", False)]):
-        return 1.0
-    elif ss == frozenset([("A", True), ("B", True)]):
-        return 10.0
-    else:
-        raise ValueError("product error")
-                                                                   
-def phi_BC(scope_product):
-    ""
-    ss = frozenset(scope_product)
-    if ss == frozenset([("B", False), ("C", False)]):
-        return 100.0
-    elif ss == frozenset([("B", False), ("C", True)]):
-        return 1.0
-    elif ss == frozenset([("B", True), ("C", False)]):
-        return 1.0
-    elif ss == frozenset([("B", True), ("C", True)]):
-        return 100.0
-    else:
-        raise ValueError("product error")
-                                                                   
-def phi_CD(scope_product):
-    ""
-    ss = frozenset(scope_product)
-    if ss == frozenset([("C", False), ("D", False)]):
-        return 1.0
-    elif ss == frozenset([("C", False), ("D", True)]):
-        return 100.0
-    elif ss == frozenset([("C", True), ("D", False)]):
-        return 100.0
-    elif ss == frozenset([("C", True), ("D", True)]):
-        return 1.0
-    else:
-        raise ValueError("product error")
[…]
