+++ docs/mdpages/usage.md
@@ -17,87 +17,87 @@ Usage:
 
 \code{.py}
 
->>> from gmodels.pgmodel import PGModel
->>> from gmodels.gtypes.edge import Edge, EdgeType
->>> from gmodels.factor import Factor
->>> from gmodels.randomvariable import NumCatRVariable
->>> # Example adapted from Darwiche 2009, p. 140
->>> idata = {
->>>             "a": {"outcome-values": [True, False]},
->>>             "b": {"outcome-values": [True, False]},
->>>             "c": {"outcome-values": [True, False]},
->>>         }
->>> a = NumCatRVariable(
->>>             node_id="a", input_data=idata["a"], distribution=lambda x: 0.6 if x else 0.4
->>>         )
->>> b = NumCatRVariable(
->>>     node_id="b", input_data=idata["b"], distribution=lambda x: 0.5 if x else 0.5
->>> )
->>> c = NumCatRVariable(
->>>     node_id="c", input_data=idata["c"], distribution=lambda x: 0.5 if x else 0.5
->>> )
->>> ab = Edge(
->>>     edge_id="ab",
->>>     edge_type=EdgeType.UNDIRECTED,
->>>     start_node=a,
->>>     end_node=b,
->>> )
->>> bc = Edge(
->>>     edge_id="bc",
->>>     edge_type=EdgeType.UNDIRECTED,
->>>     start_node=b,
->>>     end_node=c,
->>> )
->>> def phi_ba(scope_product):
->>>     ""
->>>     ss = set(scope_product)
->>>     if ss == set([("a", True), ("b", True)]):
->>>         return 0.9
->>>     elif ss == set([("a", True), ("b", False)]):
->>>         return 0.1
->>>     elif ss == set([("a", False), ("b", True)]):
->>>         return 0.2
->>>     elif ss == set([("a", False), ("b", False)]):
->>>         return 0.8
->>>     else:
->>>         raise ValueError("product error")
+from gmodels.pgmodel import PGModel
+from gmodels.gtypes.edge import Edge, EdgeType
+from gmodels.factor import Factor
+from gmodels.randomvariable import NumCatRVariable
+# Example adapted from Darwiche 2009, p. 140
+idata = {
+            "a": {"outcome-values": [True, False]},
+            "b": {"outcome-values": [True, False]},
+            "c": {"outcome-values": [True, False]},
+        }
+a = NumCatRVariable(
+            node_id="a", input_data=idata["a"], distribution=lambda x: 0.6 if x else 0.4
+        )
+b = NumCatRVariable(
+    node_id="b", input_data=idata["b"], distribution=lambda x: 0.5 if x else 0.5
+)
+c = NumCatRVariable(
+    node_id="c", input_data=idata["c"], distribution=lambda x: 0.5 if x else 0.5
+)
+ab = Edge(
+    edge_id="ab",
+    edge_type=EdgeType.UNDIRECTED,
+    start_node=a,
+    end_node=b,
+)
+bc = Edge(
+    edge_id="bc",
+    edge_type=EdgeType.UNDIRECTED,
+    start_node=b,
+    end_node=c,
+)
+def phi_ba(scope_product):
+    ""
+    ss = set(scope_product)
+    if ss == set([("a", True), ("b", True)]):
+        return 0.9
+    elif ss == set([("a", True), ("b", False)]):
+        return 0.1
+    elif ss == set([("a", False), ("b", True)]):
+        return 0.2
+    elif ss == set([("a", False), ("b", False)]):
+        return 0.8
+    else:
+        raise ValueError("product error")
 
->>> def phi_cb(scope_product):
->>>     ""
->>>     ss = set(scope_product)
[…]
