### 2021-05-01 6c193de
New tests and usage examples done
New tests for conditional random fields written.
The usage example is added to documentation.
All tests passes as well.
With CRFs gone most if not all the models are now tested.
LWF chains require a little more attention.
Continuous variables should be added
