### 2021-05-01 d44d150
New tests and usage examples done
New tests for conditional random fields written.
The usage example is added to documentation.
All tests passes as well.
With CRFs gone most if not all the models are now tested.
LWF chains require a little more attention.
Continuous variables should be added
