+++ gmodels/markov.py
@@ -3,11 +3,12 @@ Markov network
 """
 from gmodels.gtypes.undigraph import UndiGraph
 from gmodels.gtypes.edge import Edge
-from gmodels.randomvariable import NumCatRVariable
+from gmodels.randomvariable import NumCatRVariable, RandomVariable, NumericValue
 from gmodels.factor import Factor
 from gmodels.pgmodel import PGModel
 from typing import Set, Optional, Tuple
 from uuid import uuid4
+import pdb
 
 
 class MarkovNetwork(PGModel, UndiGraph):
@@ -20,7 +21,9 @@ class MarkovNetwork(PGModel, UndiGraph):
         data={},
     ):
         """!
-        Markov Random Field
+        \brief Markov Random Field implementation
+
+        For parameters \see PGModel
 
         \code{.py}
 
@@ -154,6 +157,9 @@ class MarkovNetwork(PGModel, UndiGraph):
         """!
         \brief Make a markov network from undirected graph
 
+        \throws ValueError If nodes are not an instance of a random variable,
+        we raise a value error.
+
         Unless it is specified we assume that edges indicate a joint
         distribution
 
@@ -201,6 +207,9 @@ class MarkovNetwork(PGModel, UndiGraph):
 
         \endcode
         """
+        for n in udi.nodes():
+            if not isinstance(n, RandomVariable):
+                raise ValueError("Nodes are not an instance of random variable")
         fs: Set[Factor] = set()
         maximal_cliques = udi.find_maximal_cliques()
         for clique in maximal_cliques:
@@ -233,13 +242,158 @@ class ConditionalRandomField(MarkovNetwork):
         data={},
     ):
         """!
-        CRF constructor
+        \brief Conditional Random Field
+
+
+        Several traits distinguish Conditional Random Fields (CRFs) from Markov
+        Networks.
+        Formally Conditional random fields are defined by Koller, Friedman 2009 p. 143
+        as: <blockquote> an undirected graph whose nodes correspond to a union of a set
+        of observed random variables X, and a set of target random variables Y; the
+        network is annotated with a set of factors \f$\phi_1(D_1), \dots, \phi_i(D_i),
+        \dots, \phi_m(D_m)\f$ such that \f$D_i \not \subset X\f$.
+        </blockquote>
+        The network encodes a conditional distribution between target and observed
+        variables.
+
+        The purpose of CRFs is best described by Sucar 2015, p. 92:
+        <blockquote>
+        Conditional models are used to label an observation sequence X by
+        selecting the label sequence Y that maximizes the conditional
+        probability P(Y|X). The conditional nature of such models means that no
+        effort is wasted on modeling observations, and one is free from having
+        to make unnecessary independence assumptions.
+        </blockquote>
+
+        \throws ValueError We raise a value error if a factor's scope is a
+        subset of observed variables
+
+        \param observed_vars observed random variables. These variables must be
+        different than target variables. They are allowed to have edges between
+        them. The model is conditioned on these variables
+
+        \param target_vars target random variables.
+
+        \param factors factors that encode the conditional distribution of the
+        model
+
+        \code{.py}
+
+        >>> idata = {"A": {"outcome-values": [True, False]}}
+
[…]
