### 2021-05-01 D-K-E/graphical-models#3 @D-K-E
@ankurankan I added some new tests and added a usage examples under `Related Pages` section of the documentation.
As of now, apart from the standard `BayesianNetwork` all sub classes of `PGModel` that is all that can be subjected to inference are documented.

Most my examples are toy examples taken from text books. I'll try to add more real world examples when I have a bit more time.
How does that sound ?
