### 2021-05-02 e3bacfe
new tests and bayesian usage example added
