+++ gmodels/bayesian.py
@@ -7,7 +7,7 @@ from gmodels.gtypes.edge import Edge
 from gmodels.randomvariable import NumCatRVariable
 from gmodels.factor import Factor
 from gmodels.pgmodel import PGModel
-from typing import Set
+from typing import Set, Optional, Callable
 from uuid import uuid4
 
 
@@ -24,11 +24,140 @@ class BayesianNetwork(PGModel, DiGraph):
         factors: Set[Factor],
         data={},
     ):
-        ""
+        """!
+        \see PGModel for parameters
+
+        Simple bayesian network implementation where edges are checked for
+        being directed
+
+        \code{.py}
+
+        >>> idata = {"outcome-values": [True, False]}
+
+        >>> C = NumCatRVariable(
+        >>>     node_id="C", input_data=idata, distribution=lambda x: 0.5
+        >>> )
+        >>> E = NumCatRVariable(
+        >>>     node_id="E", input_data=idata, distribution=lambda x: 0.5
+        >>> )
+        >>> F = NumCatRVariable(
+        >>>     node_id="F", input_data=idata, distribution=lambda x: 0.5
+        >>> )
+        >>> D = NumCatRVariable(
+        >>>     node_id="D", input_data=idata, distribution=lambda x: 0.5
+        >>> )
+        >>> CE = Edge(
+        >>>   edge_id="CE",
+        >>>   start_node=C,
+        >>>   end_node=E,
+        >>>   edge_type=EdgeType.DIRECTED,
+        >>> )
+        >>> ED = Edge(
+        >>>     edge_id="ED",
+        >>>     start_node=E,
+        >>>     end_node=D,
+        >>>     edge_type=EdgeType.DIRECTED,
+        >>> )
+        >>> EF = Edge(
+        >>>     edge_id="EF",
+        >>>     start_node=E,
+        >>>     end_node=F,
+        >>>     edge_type=EdgeType.DIRECTED,
+        >>> )
+
+        >>> def phi_c(scope_product):
+        >>>     ss = set(scope_product)
+        >>>    if ss == set([("C", True)]):
+        >>>        return 0.8
+        >>>     elif ss == set([("C", False)]):
+        >>>         return 0.2
+        >>>     else:
+        >>>         raise ValueError("scope product unknown")
+
+        >>> def phi_ec(scope_product):
+        >>>     ss = set(scope_product)
+        >>>     if ss == set([("C", True), ("E", True)]):
+        >>>         return 0.9
+        >>>     elif ss == set([("C", True), ("E", False)]):
+        >>>         return 0.1
+        >>>   elif ss == set([("C", False), ("E", True)]):
+        >>>       return 0.7
+        >>>   elif ss == set([("C", False), ("E", False)]):
+        >>>       return 0.3
+        >>>   else:
+        >>>        raise ValueError("scope product unknown")
+
+        >>> def phi_fe(scope_product):
+        >>>   ss = set(scope_product)
+        >>>   if ss == set([("E", True), ("F", True)]):
+        >>>       return 0.9
+        >>>   elif ss == set([("E", True), ("F", False)]):
+        >>>       return 0.1
+        >>>   elif ss == set([("E", False), ("F", True)]):
+        >>>         return 0.5
+        >>>     elif ss == set([("E", False), ("F", False)]):
+        >>>         return 0.5
+        >>>     else:
+        >>>         raise ValueError("scope product unknown")
+
+        >>> def phi_de(scope_product):
+        >>>     ss = set(scope_product)
+        >>>   if ss == set([("E", True), ("D", True)]):
+        >>>       return 0.7
[…]
