### 2021-05-02 ff30e1a
new tests and bayesian usage example added
