+++ test/test_bayesian.py
@@ -10,6 +10,7 @@ from gmodels.randomvariable import NumCatRVariable
 from uuid import uuid4
 
 import unittest
+import pdb
 
 
 class BayesianNetworkTest(unittest.TestCase):
@@ -30,17 +31,17 @@ class BayesianNetworkTest(unittest.TestCase):
         self.rain = NumCatRVariable(
             input_data=idata["rain"],
             node_id="rain",
-            distribution=lambda x: 0.2 if x is True else 0.8,
+            marginal_distribution=lambda x: 0.2 if x is True else 0.8,
         )
         self.sprink = NumCatRVariable(
             node_id="sprink",
             input_data=idata["sprink"],
-            distribution=lambda x: 0.6 if x is True else 0.4,
+            marginal_distribution=lambda x: 0.6 if x is True else 0.4,
         )
         self.wet = NumCatRVariable(
             node_id="wet",
             input_data=idata["wet"],
-            distribution=lambda x: 0.7 if x is True else 0.3,
+            marginal_distribution=lambda x: 0.7 if x is True else 0.3,
         )
         self.rain_wet = Edge(
             edge_id="rain_wet",
@@ -75,8 +76,8 @@ class BayesianNetworkTest(unittest.TestCase):
             else:
                 raise ValueError("unknown product")
 
-        self.rain_sprink_f = Factor.from_conditional_vars(
-            X_i=self.sprink, Pa_Xi=set([self.rain]), fn=sprink_rain_factor
+        self.rain_sprink_f = Factor.from_scope_variables_with_fn(
+            svars=set([self.rain, self.sprink]), fn=sprink_rain_factor
         )
 
         def grass_wet_factor(scope_product):
@@ -101,8 +102,8 @@ class BayesianNetworkTest(unittest.TestCase):
             else:
                 raise ValueError("unknown product")
 
-        self.grass_wet_f = Factor.from_conditional_vars(
-            X_i=self.wet, Pa_Xi=set([self.rain, self.sprink]), fn=grass_wet_factor
+        self.grass_wet_f = Factor.from_scope_variables_with_fn(
+            svars=set([self.rain, self.sprink, self.wet]), fn=grass_wet_factor
         )
 
         self.bayes = BayesianNetwork(
@@ -122,17 +123,17 @@ class BayesianNetworkTest(unittest.TestCase):
         self.EarthquakeN = NumCatRVariable(
             input_data=idata["earthquake"],
             node_id="EarthquakeN",
-            distribution=lambda x: 0.1 if x is True else 0.9,
+            marginal_distribution=lambda x: 0.1 if x is True else 0.9,
         )
         self.BurglaryN = NumCatRVariable(
             input_data=idata["burglary"],
             node_id="BurglaryN",
-            distribution=lambda x: 0.2 if x is True else 0.8,
+            marginal_distribution=lambda x: 0.2 if x is True else 0.8,
         )
         self.AlarmN = NumCatRVariable(
             input_data=idata["alarm"],
             node_id="AlarmN",
-            distribution=lambda x: 0.2442 if x is True else 0.7558,
+            marginal_distribution=lambda x: 0.2442 if x is True else 0.7558,
         )
         self.burglar_alarm = Edge(
             edge_id="burglar_alarm",
@@ -150,16 +151,16 @@ class BayesianNetworkTest(unittest.TestCase):
         idata = {"outcome-values": [True, False]}
 
         self.C = NumCatRVariable(
-            node_id="C", input_data=idata, distribution=lambda x: 0.5
+            node_id="C", input_data=idata, marginal_distribution=lambda x: 0.5
         )
         self.E = NumCatRVariable(
-            node_id="E", input_data=idata, distribution=lambda x: 0.5
[…]

[…]
