### 2021-05-02 aa91859
added tests to bayesian and usage examples
