### 2021-05-03 87399ab
documentation and usage example update, markov test
