### 2021-05-03 e044008
documentation and usage example update, markov test
