### 2021-05-03 #3115 @D-K-E
I have updated documentation with usage examples. Now all the probablistic graphical models that can be inferred have a usage example. All of them have a unittest as well.

I'll try to improve the documentation on the intermediary objects.
