### 2021-05-04 D-K-E/graphical-models#3 @ankurankan
@D-K-E The examples look good, but I am getting errors in running all of them (except the first one). Could you please check?
