### 2021-05-06 5d66faa
Corrected usage examples added notebooks
Added jupyter notebooks for demonstrating the use of the package.
The documentation is also corrected.
