+++ docs/mdpages/usage.md
@@ -127,20 +127,30 @@ Usage:
 
 \code{.py}
 
+# import necessary parts
+
+from gmodels.bayesian import BayesianNetwork
+from gmodels.gtypes.edge import Edge, EdgeType
+from gmodels.factor import Factor
+from gmodels.randomvariable import NumCatRVariable
+
+# data and nodes
 idata = {"outcome-values": [True, False]}
                                                               
 C = NumCatRVariable(
-    node_id="C", input_data=idata, distribution=lambda x: 0.5
+    node_id="C", input_data=idata, marginal_distribution=lambda x: 0.5
 )
 E = NumCatRVariable(
-    node_id="E", input_data=idata, distribution=lambda x: 0.5
+    node_id="E", input_data=idata, marginal_distribution=lambda x: 0.5
 )
 F = NumCatRVariable(
-    node_id="F", input_data=idata, distribution=lambda x: 0.5
+    node_id="F", input_data=idata, marginal_distribution=lambda x: 0.5
 )
 D = NumCatRVariable(
-    node_id="D", input_data=idata, distribution=lambda x: 0.5
+    node_id="D", input_data=idata, marginal_distribution=lambda x: 0.5
 )
+
+# edges
 CE = Edge(
   edge_id="CE",
   start_node=C,
@@ -159,11 +169,13 @@ EF = Edge(
     end_node=F,
     edge_type=EdgeType.DIRECTED,
 )
-                                                              
+
+# define factor functions
+
 def phi_c(scope_product):
     ss = set(scope_product)
-   if ss == set([("C", True)]):
-       return 0.8
+    if ss == set([("C", True)]):
+        return 0.8
     elif ss == set([("C", False)]):
         return 0.2
     else:
@@ -175,39 +187,41 @@ def phi_ec(scope_product):
         return 0.9
     elif ss == set([("C", True), ("E", False)]):
         return 0.1
-  elif ss == set([("C", False), ("E", True)]):
-      return 0.7
-  elif ss == set([("C", False), ("E", False)]):
-      return 0.3
-  else:
-       raise ValueError("scope product unknown")
+    elif ss == set([("C", False), ("E", True)]):
+        return 0.7
+    elif ss == set([("C", False), ("E", False)]):
+        return 0.3
+    else:
+        raise ValueError("scope product unknown")
                                                               
 def phi_fe(scope_product):
-  ss = set(scope_product)
-  if ss == set([("E", True), ("F", True)]):
-      return 0.9
-  elif ss == set([("E", True), ("F", False)]):
-      return 0.1
-  elif ss == set([("E", False), ("F", True)]):
+    ss = set(scope_product)
+    if ss == set([("E", True), ("F", True)]):
+        return 0.9
+    elif ss == set([("E", True), ("F", False)]):
+        return 0.1
+    elif ss == set([("E", False), ("F", True)]):
         return 0.5
     elif ss == set([("E", False), ("F", False)]):
         return 0.5
     else:
         raise ValueError("scope product unknown")
-                                                              
+                                            
 def phi_de(scope_product):
     ss = set(scope_product)
-  if ss == set([("E", True), ("D", True)]):
-      return 0.7
[…]
