### 2021-05-06 D-K-E/graphical-models#3 @D-K-E
@ankurankan All right, they all should have been fixed now, both in the documentation and the newly added jupyet notebooks
