### 2021-05-10 7904df0
new tests for graphs and updated documentation
