### 2021-05-12 fed8316
some JOSS edits
