+++ paper/paper.md
@@ -10,7 +10,7 @@ authors:
   affiliation: 1
   orcid: 0000-0002-1552-8938
 affiliations:
-- name: EPHE, PSL
+- name: École Pratique des Hautes Études, Université PSL, Paris, France
   index: 1
 date: 02 February 2021
 bibliography: paper.bib
@@ -18,62 +18,54 @@ bibliography: paper.bib
 
 # Summary
 
-Probabilistic Graphical Models (PGMs) are a marriage between Graphs as in
-Graph Theory, and Probability as in statistics and probability theory and are
-widely used in many fields. We noticed that most existing PGM
-libraries implement PGMs in a way that ignores their graphical nature.
+Probabilistic Graphical Models (PGMs) are a marriage between "graphs" from
+graph theory and "probability" from statistics and probability theory. While PGMs are
+widely used in many fields, we noticed that most existing PGM
+libraries are implement in a way that doesn't take full advantage of their graphical nature.
 `PyGModels`' value proposition is that it faithfully implements the graphical
 nature of PGMs, thereby giving `PyGModels`' instantiated objects both
 graph-theoretical and statistical properties, which allows users to explore
-and test inference algorithms that are rooted in both graph theory or
-statistics. `PyGModels` also implements several algorithms of interest on a
-LWF chain graphs, also known as mixed graphs.
+and test inference algorithms that are rooted in graph theory or
+statistics. `PyGModels` also implements several algorithms of interest on Lauritzen-Wermuth-Frydenberg
+(LWF) chain graphs, also known as mixed graphs.
 
 # Statement of Need
 
-Though the students of computer science or statistics might find a pedagogical
+Though the students of computer science or statistics might find pedagogical
 value going through source code along with a textbook on probabilistic
-graphical models (something like Sucar [see @Sucar_2015] or Cowell [see
-@Cowell_2005] or Koller and Friedman [see @Koller_Friedman_2009]), we believe
-that the value proposition of `PyGModels` speaks mostly to researchers.
-Let us try to demonstrate the need for `PyGModels` by a use case. 
+graphical models [for example @Cowell_2005; @Koller_Friedman_2009; @Sucar_2015],
+`PyGModels` is mainly targeted at researchers.
+Let us demonstrate the need for `PyGModels` with a use case.
+Given a set of categorical random variables in the form of a function
+specified by a probability distribution, a set of edges that encode a
+certain independence assumption over these random variables, and a set
+of factors that factorizes this probability distribution over the
+graph, `PyGModels` is designed with the following use cases in mind:
 
-One has a set of categorical random variables in the form of a function
-specified by a probability distribution. One has a set of edges that encode a
-certain independence assumption over her random variables, and one has a set
-of factors, that factorizes a certain probability distribution over her entire
-graph. Given these, `PyGModels` might solve two major issues for the
-researcher or the student alike:
+1. computation of the posterior probability distribution or most probable explanation
+  conditioned on evidence, and
 
-1. Compute posterior probability distribution or most probable explanation
-  given certain evidence.
+2. the development of new inference algorithms.
 
-2. Provide a basis for creating new algorithms of inference.
+The real forte of `PyGModels` is its support for implementing new algorithms
+due to its lightweight nature and its direct implementation of
+statistical and graph theoretic features in the same base class.
+We mostly follow @Koller_Friedman_2009 for statistical conventions, definitions,
+and inference algorithms. For graph theoretic conventions, we follow
+@Diestel_2017, with algorithms from @Erciyes_2018 and @Even_Guy_Even_2012.
+Throughout the code, exact pages for algorithmic references are cited in the
+docstrings for relevant functions.
 
[…]
