### 2021-05-12 #3115 @D-K-E
@dfm All right I think I did what you said:
The new version is `v0.1.0`.
The release is called `JOSS version`.
 DOI: 10.5281/zenodo.4751740

The title and the author are the same. Affiliation is okay as well. The manuscript is all good as well.
